Convert a data.frame to a Matrix in Plate Format

Extracts row and column information from the column specified by well_col (defaults to 'well') and uses this to create a matrix that contains values from the values_from column.

well_df_to_mat(df, values_from, well_col = "well", plate = NULL)

Arguments

df

A dataframe containing at least a column called "well" that contains well IDs.

values_from

Column that contains values to populate matrix.

well_col

Column containing the well IDs to extract row and column information from.

plate

Size of the plate, to override the auto-detected plate size.

Value

A matrix containing the values that were present in the data frame in the values_from column

Examples

library(wellr)

# pivot a long data frame into a matrix based on the row and col columns,
# with values from a given column
plate <- well_plate(8, 12)
plate$value <- seq(nrow(plate))
well_df_to_mat(plate, values_from = "value")
#>    row col well value
#> 1    1   1  A01     1
#> 2    1   2  A02     2
#> 3    1   3  A03     3
#> 4    1   4  A04     4
#> 5    1   5  A05     5
#> 6    1   6  A06     6
#> 7    1   7  A07     7
#> 8    1   8  A08     8
#> 9    1   9  A09     9
#> 10   1  10  A10    10
#> 11   1  11  A11    11
#> 12   1  12  A12    12
#> 13   2   1  B01    13
#> 14   2   2  B02    14
#> 15   2   3  B03    15
#> 16   2   4  B04    16
#> 17   2   5  B05    17
#> 18   2   6  B06    18
#> 19   2   7  B07    19
#> 20   2   8  B08    20
#> 21   2   9  B09    21
#> 22   2  10  B10    22
#> 23   2  11  B11    23
#> 24   2  12  B12    24
#> 25   3   1  C01    25
#> 26   3   2  C02    26
#> 27   3   3  C03    27
#> 28   3   4  C04    28
#> 29   3   5  C05    29
#> 30   3   6  C06    30
#> 31   3   7  C07    31
#> 32   3   8  C08    32
#> 33   3   9  C09    33
#> 34   3  10  C10    34
#> 35   3  11  C11    35
#> 36   3  12  C12    36
#> 37   4   1  D01    37
#> 38   4   2  D02    38
#> 39   4   3  D03    39
#> 40   4   4  D04    40
#> 41   4   5  D05    41
#> 42   4   6  D06    42
#> 43   4   7  D07    43
#> 44   4   8  D08    44
#> 45   4   9  D09    45
#> 46   4  10  D10    46
#> 47   4  11  D11    47
#> 48   4  12  D12    48
#> 49   5   1  E01    49
#> 50   5   2  E02    50
#> 51   5   3  E03    51
#> 52   5   4  E04    52
#> 53   5   5  E05    53
#> 54   5   6  E06    54
#> 55   5   7  E07    55
#> 56   5   8  E08    56
#> 57   5   9  E09    57
#> 58   5  10  E10    58
#> 59   5  11  E11    59
#> 60   5  12  E12    60
#> 61   6   1  F01    61
#> 62   6   2  F02    62
#> 63   6   3  F03    63
#> 64   6   4  F04    64
#> 65   6   5  F05    65
#> 66   6   6  F06    66
#> 67   6   7  F07    67
#> 68   6   8  F08    68
#> 69   6   9  F09    69
#> 70   6  10  F10    70
#> 71   6  11  F11    71
#> 72   6  12  F12    72
#> 73   7   1  G01    73
#> 74   7   2  G02    74
#> 75   7   3  G03    75
#> 76   7   4  G04    76
#> 77   7   5  G05    77
#> 78   7   6  G06    78
#> 79   7   7  G07    79
#> 80   7   8  G08    80
#> 81   7   9  G09    81
#> 82   7  10  G10    82
#> 83   7  11  G11    83
#> 84   7  12  G12    84
#> 85   8   1  H01    85
#> 86   8   2  H02    86
#> 87   8   3  H03    87
#> 88   8   4  H04    88
#> 89   8   5  H05    89
#> 90   8   6  H06    90
#> 91   8   7  H07    91
#> 92   8   8  H08    92
#> 93   8   9  H09    93
#> 94   8  10  H10    94
#> 95   8  11  H11    95
#> 96   8  12  H12    96
#>    row col well value
#> 1    1   1  A01     1
#> 2    1   2  A02     2
#> 3    1   3  A03     3
#> 4    1   4  A04     4
#> 5    1   5  A05     5
#> 6    1   6  A06     6
#> 7    1   7  A07     7
#> 8    1   8  A08     8
#> 9    1   9  A09     9
#> 10   1  10  A10    10
#> 11   1  11  A11    11
#> 12   1  12  A12    12
#> 13   2   1  B01    13
#> 14   2   2  B02    14
#> 15   2   3  B03    15
#> 16   2   4  B04    16
#> 17   2   5  B05    17
#> 18   2   6  B06    18
#> 19   2   7  B07    19
#> 20   2   8  B08    20
#> 21   2   9  B09    21
#> 22   2  10  B10    22
#> 23   2  11  B11    23
#> 24   2  12  B12    24
#> 25   3   1  C01    25
#> 26   3   2  C02    26
#> 27   3   3  C03    27
#> 28   3   4  C04    28
#> 29   3   5  C05    29
#> 30   3   6  C06    30
#> 31   3   7  C07    31
#> 32   3   8  C08    32
#> 33   3   9  C09    33
#> 34   3  10  C10    34
#> 35   3  11  C11    35
#> 36   3  12  C12    36
#> 37   4   1  D01    37
#> 38   4   2  D02    38
#> 39   4   3  D03    39
#> 40   4   4  D04    40
#> 41   4   5  D05    41
#> 42   4   6  D06    42
#> 43   4   7  D07    43
#> 44   4   8  D08    44
#> 45   4   9  D09    45
#> 46   4  10  D10    46
#> 47   4  11  D11    47
#> 48   4  12  D12    48
#> 49   5   1  E01    49
#> 50   5   2  E02    50
#> 51   5   3  E03    51
#> 52   5   4  E04    52
#> 53   5   5  E05    53
#> 54   5   6  E06    54
#> 55   5   7  E07    55
#> 56   5   8  E08    56
#> 57   5   9  E09    57
#> 58   5  10  E10    58
#> 59   5  11  E11    59
#> 60   5  12  E12    60
#> 61   6   1  F01    61
#> 62   6   2  F02    62
#> 63   6   3  F03    63
#> 64   6   4  F04    64
#> 65   6   5  F05    65
#> 66   6   6  F06    66
#> 67   6   7  F07    67
#> 68   6   8  F08    68
#> 69   6   9  F09    69
#> 70   6  10  F10    70
#> 71   6  11  F11    71
#> 72   6  12  F12    72
#> 73   7   1  G01    73
#> 74   7   2  G02    74
#> 75   7   3  G03    75
#> 76   7   4  G04    76
#> 77   7   5  G05    77
#> 78   7   6  G06    78
#> 79   7   7  G07    79
#> 80   7   8  G08    80
#> 81   7   9  G09    81
#> 82   7  10  G10    82
#> 83   7  11  G11    83
#> 84   7  12  G12    84
#> 85   8   1  H01    85
#> 86   8   2  H02    86
#> 87   8   3  H03    87
#> 88   8   4  H04    88
#> 89   8   5  H05    89
#> 90   8   6  H06    90
#> 91   8   7  H07    91
#> 92   8   8  H08    92
#> 93   8   9  H09    93
#> 94   8  10  H10    94
#> 95   8  11  H11    95
#> 96   8  12  H12    96
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12]
#> [1,]    1    2    3    4    5    6    7    8    9    10    11    12
#> [2,]   13   14   15   16   17   18   19   20   21    22    23    24
#> [3,]   25   26   27   28   29   30   31   32   33    34    35    36
#> [4,]   37   38   39   40   41   42   43   44   45    46    47    48
#> [5,]   49   50   51   52   53   54   55   56   57    58    59    60
#> [6,]   61   62   63   64   65   66   67   68   69    70    71    72
#> [7,]   73   74   75   76   77   78   79   80   81    82    83    84
#> [8,]   85   86   87   88   89   90   91   92   93    94    95    96

# the function also handles missing values from a plate, filling with with NA
# if certain wells are missing
plate <- well_plate(7, 11)
plate$value <- seq(nrow(plate))
well_df_to_mat(plate, values_from = "value", plate = 384)
#>    row col well value
#> 1    1   1  A01     1
#> 2    1   2  A02     2
#> 3    1   3  A03     3
#> 4    1   4  A04     4
#> 5    1   5  A05     5
#> 6    1   6  A06     6
#> 7    1   7  A07     7
#> 8    1   8  A08     8
#> 9    1   9  A09     9
#> 10   1  10  A10    10
#> 11   1  11  A11    11
#> 12   2   1  B01    12
#> 13   2   2  B02    13
#> 14   2   3  B03    14
#> 15   2   4  B04    15
#> 16   2   5  B05    16
#> 17   2   6  B06    17
#> 18   2   7  B07    18
#> 19   2   8  B08    19
#> 20   2   9  B09    20
#> 21   2  10  B10    21
#> 22   2  11  B11    22
#> 23   3   1  C01    23
#> 24   3   2  C02    24
#> 25   3   3  C03    25
#> 26   3   4  C04    26
#> 27   3   5  C05    27
#> 28   3   6  C06    28
#> 29   3   7  C07    29
#> 30   3   8  C08    30
#> 31   3   9  C09    31
#> 32   3  10  C10    32
#> 33   3  11  C11    33
#> 34   4   1  D01    34
#> 35   4   2  D02    35
#> 36   4   3  D03    36
#> 37   4   4  D04    37
#> 38   4   5  D05    38
#> 39   4   6  D06    39
#> 40   4   7  D07    40
#> 41   4   8  D08    41
#> 42   4   9  D09    42
#> 43   4  10  D10    43
#> 44   4  11  D11    44
#> 45   5   1  E01    45
#> 46   5   2  E02    46
#> 47   5   3  E03    47
#> 48   5   4  E04    48
#> 49   5   5  E05    49
#> 50   5   6  E06    50
#> 51   5   7  E07    51
#> 52   5   8  E08    52
#> 53   5   9  E09    53
#> 54   5  10  E10    54
#> 55   5  11  E11    55
#> 56   6   1  F01    56
#> 57   6   2  F02    57
#> 58   6   3  F03    58
#> 59   6   4  F04    59
#> 60   6   5  F05    60
#> 61   6   6  F06    61
#> 62   6   7  F07    62
#> 63   6   8  F08    63
#> 64   6   9  F09    64
#> 65   6  10  F10    65
#> 66   6  11  F11    66
#> 67   7   1  G01    67
#> 68   7   2  G02    68
#> 69   7   3  G03    69
#> 70   7   4  G04    70
#> 71   7   5  G05    71
#> 72   7   6  G06    72
#> 73   7   7  G07    73
#> 74   7   8  G08    74
#> 75   7   9  G09    75
#> 76   7  10  G10    76
#> 77   7  11  G11    77
#>    row col well value
#> 1    1   1  A01     1
#> 2    1   2  A02     2
#> 3    1   3  A03     3
#> 4    1   4  A04     4
#> 5    1   5  A05     5
#> 6    1   6  A06     6
#> 7    1   7  A07     7
#> 8    1   8  A08     8
#> 9    1   9  A09     9
#> 10   1  10  A10    10
#> 11   1  11  A11    11
#> 12   2   1  B01    12
#> 13   2   2  B02    13
#> 14   2   3  B03    14
#> 15   2   4  B04    15
#> 16   2   5  B05    16
#> 17   2   6  B06    17
#> 18   2   7  B07    18
#> 19   2   8  B08    19
#> 20   2   9  B09    20
#> 21   2  10  B10    21
#> 22   2  11  B11    22
#> 23   3   1  C01    23
#> 24   3   2  C02    24
#> 25   3   3  C03    25
#> 26   3   4  C04    26
#> 27   3   5  C05    27
#> 28   3   6  C06    28
#> 29   3   7  C07    29
#> 30   3   8  C08    30
#> 31   3   9  C09    31
#> 32   3  10  C10    32
#> 33   3  11  C11    33
#> 34   4   1  D01    34
#> 35   4   2  D02    35
#> 36   4   3  D03    36
#> 37   4   4  D04    37
#> 38   4   5  D05    38
#> 39   4   6  D06    39
#> 40   4   7  D07    40
#> 41   4   8  D08    41
#> 42   4   9  D09    42
#> 43   4  10  D10    43
#> 44   4  11  D11    44
#> 45   5   1  E01    45
#> 46   5   2  E02    46
#> 47   5   3  E03    47
#> 48   5   4  E04    48
#> 49   5   5  E05    49
#> 50   5   6  E06    50
#> 51   5   7  E07    51
#> 52   5   8  E08    52
#> 53   5   9  E09    53
#> 54   5  10  E10    54
#> 55   5  11  E11    55
#> 56   6   1  F01    56
#> 57   6   2  F02    57
#> 58   6   3  F03    58
#> 59   6   4  F04    59
#> 60   6   5  F05    60
#> 61   6   6  F06    61
#> 62   6   7  F07    62
#> 63   6   8  F08    63
#> 64   6   9  F09    64
#> 65   6  10  F10    65
#> 66   6  11  F11    66
#> 67   7   1  G01    67
#> 68   7   2  G02    68
#> 69   7   3  G03    69
#> 70   7   4  G04    70
#> 71   7   5  G05    71
#> 72   7   6  G06    72
#> 73   7   7  G07    73
#> 74   7   8  G08    74
#> 75   7   9  G09    75
#> 76   7  10  G10    76
#> 77   7  11  G11    77
#> Warning: data length [77] is not a sub-multiple or multiple of the number of rows [8]
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12]
#> [1,]    1    2    3    4    5    6    7    8    9    10    11    12
#> [2,]   13   14   15   16   17   18   19   20   21    22    23    24
#> [3,]   25   26   27   28   29   30   31   32   33    34    35    36
#> [4,]   37   38   39   40   41   42   43   44   45    46    47    48
#> [5,]   49   50   51   52   53   54   55   56   57    58    59    60
#> [6,]   61   62   63   64   65   66   67   68   69    70    71    72
#> [7,]   73   74   75   76   77    1    2    3    4     5     6     7
#> [8,]    8    9   10   11   12   13   14   15   16    17    18    19