Convert a DataFrame to a Multi-Frame Matrix

Converts a DataFrame to a matrix where each row is a single time point, and each column is a single well. The wells are in index order, indexing across the rows so the first values are all from row A.

well_df_to_mat_frames(data, value, frame, well)

Arguments

data

Data frame or tibble to convert to a multi-frame matrix.

value

Column with value data that will be used to populate the matrix.

frame

Column with the frame or time data that will be used to create the rows of the matrix.

well

Column with the well IDs which will be used to create the columns of the matrix.

Value

a matrix where each column is a well and each row is a time point.

Examples

df_list <- lapply(1:10, function(x) {
  df <- wellr::well_plate()
  df$value <- rnorm(96)
  df$frame <- x
  df
})

df_frames <- do.call(rbind, df_list)

# convert the data frame to multi-frame matrix
mat <- well_df_to_mat_frames(
  data = df_frames,
  value = value,
  frame = frame,
  well = well
)

head(mat[, 1:14])
#>          A01       A02        A03        A04          A05          A06
#> 1  0.3494660 1.9685960 -1.6011832  2.0229239  0.390258597 -0.641057954
#> 2 -0.4014917 0.7081566  1.1345810  0.7041477 -0.073218834  0.554171186
#> 3  1.0500329 1.1831186 -0.7439858  1.3427331 -1.424253828 -1.120408022
#> 4 -0.7327416 1.3528549  1.5378449  0.6399335  1.205227133 -0.367837697
#> 5  1.3615046 0.9088443 -0.0908809 -0.1786865  0.083867367  0.434657625
#> 6  0.5937080 0.2514960 -0.3344438 -0.2331829  0.006901023 -0.009799706
#>          A07        A08        A09         A10         A11         A12
#> 1 -2.1184847 -1.0989771 -0.4163115 -0.16826799  0.35026470 -0.26541776
#> 2 -1.0889587  0.1495477  0.8063449  0.12348466 -1.20657722 -0.36595457
#> 3 -0.1032231  0.9372224  1.3038912  0.09486091 -0.09179176  0.95631744
#> 4  1.1668514 -0.3388115  0.8938242 -1.00769482  2.09248633 -0.67337782
#> 5 -0.4538102  1.6425660 -1.6112164  0.60275927 -0.63492829  0.68233440
#> 6  0.3900292 -0.0538484 -0.5378362  0.79554820 -0.56757162 -0.05242387
#>          B01        B02
#> 1 -0.9182847  0.6795997
#> 2 -1.0408133 -0.5427393
#> 3 -0.5499338 -0.4303822
#> 4  0.9817243  0.3627030
#> 5 -0.4655979 -0.6288414
#> 6  0.2019844  0.1670779