Relative Distance Between Given Well and a Reference Row

Relative Distance Between Given Well and a Reference Row

well_dis(row, col, ref_row, ref_col)

Arguments

row

Numeric vector of row numbers. Must be equal in length to col.

col

Numeric vector of col numbers. Must be equal in length to row.

ref_row

Single reference row number.

ref_col

Single reference column number.

Value

a numeric vector.

Examples

well_dis(1:8, 1:12, row_row = 5, row_col = 5)
#> Error in well_dis(1:8, 1:12, row_row = 5, row_col = 5): unused arguments (row_row = 5, row_col = 5)
well_dis(1:8, 1:12, row_row = well_to_rownum("C4"), row_col = well_to_colnum("C4"))
#> Error in well_dis(1:8, 1:12, row_row = well_to_rownum("C4"), row_col = well_to_colnum("C4")): unused arguments (row_row = well_to_rownum("C4"), row_col = well_to_colnum("C4"))