Turn a Multi-Frame Matrix into a DataFrame

Takes a matrix where each row is a time point and each column is a well and converts it to a DataFrame containing a column for the well, a column for the row, a column for the column, a column for the time and a column for the values.

well_mat_frames_to_df(matrix, value_col = "value")

Arguments

matrix

multi-frame matrix to be converted.

value_col

Name to be assigned to the values column.

Value

a tibble

Examples

df_list <- lapply(1:10, function(x) {
  df <- wellr::well_plate()
  df$value <- rnorm(96)
  df$frame <- x
  df
})

df_frames <- do.call(rbind, df_list)

# convert the data frame to multi-frame matrix
mat <- well_df_to_mat_frames(
  data = df_frames,
  values_col = "value",
  time_col = "frame"
)

head(mat[, 1:14])
#>          A01         A02       A03        A04        A05        A06        A07
#> 1 -0.6611669  0.82752350 0.8926586 -0.1315812 -0.1215391  0.5978065 -1.2159612
#> 2 -1.4710447  0.58029924 0.1029971 -1.5670145 -0.5821751 -1.1272248 -0.4354051
#> 3 -0.7353637 -0.01921787 1.2349541  1.0290353  0.6043133 -0.2845863 -0.3643448
#> 4 -0.1518555  1.12393865 0.3157451 -1.4494011 -1.2049370 -0.5014733  0.6125382
#> 5  0.5002950 -0.82864816 1.5841362  0.3115112 -0.5801438 -1.5865310  0.9490984
#> 6 -1.9214425  0.27133725 0.5477939  0.7663629  0.2795711 -1.3261578 -2.1282715
#>           A08         A09        A10        A11        A12        B01
#> 1  1.23240655 -0.02445427 -1.3790382 -0.1458323 -0.4239905 -0.4120554
#> 2 -0.83740804  0.49226214  0.6618359  0.2179504 -1.1162744 -2.4847156
#> 3  0.07670452  0.68427148  1.3007369  0.4060757  1.1581952  0.6615964
#> 4  0.00870228  1.08595149  0.9303970 -0.5300271  0.1427033 -0.4486140
#> 5  0.17267547 -0.71859871 -0.3561362  0.8064921  1.4577069  0.4757358
#> 6  0.64871585  1.49018844  0.4883163 -0.6387803 -0.4146261  0.1949777
#>          B02
#> 1 -0.9064505
#> 2 -1.6571198
#> 3 -0.2668782
#> 4 -0.1415774
#> 5  0.5201374
#> 6 -0.5019792

# convert the matrix back to a dataframe
well_mat_frames_to_df(mat)
#> # A tibble: 960 × 5
#>    well    row   col   value frames
#>    <chr> <dbl> <dbl>   <dbl> <chr> 
#>  1 A01       1     1 -0.661  1     
#>  2 A02       1     2 -0.0245 1     
#>  3 A03       1     3  0.210  1     
#>  4 A04       1     4 -0.563  1     
#>  5 A05       1     5 -0.0326 1     
#>  6 A06       1     6  0.133  1     
#>  7 A07       1     7 -0.0581 1     
#>  8 A08       1     8 -0.438  1     
#>  9 A09       1     9  1.62   1     
#> 10 A10       1    10 -0.470  1     
#> # … with 950 more rows