Reorder a Plate-Based DataFrame

Reorder a Plate-Based DataFrame

well_reorder_df(data, well_col = "well", row_col = NULL, col_col = NULL)

Arguments

data

Data frame to reorder, than contains a row column and a col column.

well_col

Column containing the well IDs.

row_col

Column containing the row letters or numbers.

col_col

Column containing the column numbers.

Value

A data.frame that has been reordered according to the row & colcolumns.

Examples

df <- well_plate(nrow = 8, ncol = 12)
df <- df[sample(1:96, 96), ]
head(df)
#> # A tibble: 6 × 3
#>     row   col well 
#>   <int> <int> <chr>
#> 1     1     7 A07  
#> 2     8     8 H08  
#> 3     7    11 G11  
#> 4     8     3 H03  
#> 5     6     2 F02  
#> 6     2    12 B12  

df <- well_reorder_df(df)
#> Warning: argument is not an atomic vector; coercing
#> Warning: argument is not an atomic vector; coercing
head(df)
#> # A tibble: 6 × 3
#>     row   col well 
#>   <dbl> <dbl> <chr>
#> 1     3    NA A07  
#> 2     3    NA H08  
#> 3     3    NA G11  
#> 4     3    NA H03  
#> 5     3    NA F02  
#> 6     3    NA B12