Convert a DataFrame to a Multi-Frame Matrix

Converts a DataFrame to a matrix where each row is a single time point, and each column is a single well. The wells are in index order, indexing across the rows so the first values are all from row A.

well_df_to_mat_frames(data, values_col, time_col, well_col = "well")

Arguments

data

A DataFrame to be converted.

values_col

The name of the column containing the values.

time_col

The name of the column containing the time values.

well_col

The name of the column with the well IDs.

Value

a matrix

Examples

df_list <- lapply(1:10, function(x) {
  df <- wellr::well_plate()
  df$value <- rnorm(96)
  df$frame <- x
  df
})

df_frames <- do.call(rbind, df_list)

# convert the data frame to multi-frame matrix
mat <- well_df_to_mat_frames(
  data = df_frames,
  values_col = "value",
  time_col = "frame"
)

head(mat[, 1:14])
#>          A01         A02        A03        A04        A05        A06
#> 1  0.9355021  0.02765721  0.9063963  0.9789476 -0.2099090  1.7249723
#> 2  0.8033157 -1.79876395  0.1267061 -1.2078974  0.9128000  0.2576928
#> 3  0.1584031  0.66718465  0.7644382  1.9282962  0.2163124  1.1159240
#> 4  0.2459807  0.27864986 -0.4277557  0.6035153 -0.8017078  0.4874981
#> 5  0.6652338  0.93764343  0.2168217 -1.9891465  0.3661007 -0.2260044
#> 6 -0.4846585  0.92110324 -0.3743180 -0.8436642  1.5785394  0.8723851
#>           A07        A08          A09        A10         A11         A12
#> 1 -1.40300958  1.8464633 -0.926368849  0.2459750 -1.11260805 -1.29681537
#> 2  0.78068092  0.7280691 -1.320916420  0.8674247 -1.06139674  0.16587728
#> 3 -0.79717602 -0.2227121 -0.557577164 -0.8736790 -0.60758833 -0.41687401
#> 4  0.67618146 -0.3969969  1.773192541 -1.0316212 -0.01913135  0.50197110
#> 5  0.05222836 -0.8846907  0.650317955 -2.4339549  0.29878470 -0.08318484
#> 6  1.13452731  0.8680041  0.002628019  0.4057660 -0.01916830 -1.08967418
#>           B01        B02
#> 1 -0.43390688 -0.7805454
#> 2  0.08595339 -0.5720296
#> 3 -0.27518879  1.1251805
#> 4  2.29383563  0.6404424
#> 5 -0.52837713 -1.2311715
#> 6 -1.45716149  1.1065305