Convert a DataFrame to a Multi-Frame Matrix

Converts a DataFrame to a matrix where each row is a single time point, and each column is a single well. The wells are in index order, indexing across the rows so the first values are all from row A.

well_df_to_mat_frames(data, values_col, time_col, well_col = "well")

Arguments

data

A DataFrame to be converted.

values_col

The name of the column containing the values.

time_col

The name of the column containing the time values.

well_col

The name of the column with the well IDs.

Value

a matrix

Examples

df_list <- lapply(1:10, function(x) {
  df <- wellr::well_plate()
  df$value <- rnorm(96)
  df$frame <- x
  df
})

df_frames <- do.call(rbind, df_list)

# convert the data frame to multi-frame matrix
mat <- well_df_to_mat_frames(
  data = df_frames,
  values_col = "value",
  time_col = "frame"
)

head(mat[, 1:14])
#>           A01          A02         A03        A04         A05         A06
#> 1 -0.06533309 -0.137954607 -0.73277536  0.4171280  0.19136816 -0.08737974
#> 2  0.24957528 -0.282965583 -0.12161361 -1.2906878  0.07673215 -0.41539753
#> 3 -1.01676775  0.008327514  0.46246224  0.9040311  0.44824201  0.14559387
#> 4 -0.34111991  1.063079278 -2.03135394 -0.4969214  2.00022760  0.04564662
#> 5 -1.13768827  0.676420416 -0.07766754  0.2926566  0.16052123  1.59074662
#> 6 -1.30287426 -0.597955913 -0.93716363 -1.9709389 -0.81012483 -0.81537891
#>           A07        A08         A09        A10        A11        A12
#> 1  2.28911276  1.0340079  0.78490057 -0.1660030 -1.6473956 -0.4915235
#> 2  0.06001251  0.3209598  0.95904724 -1.1860443  1.2552197 -0.8211074
#> 3  0.55290436 -0.2948304 -0.68438054 -0.3376243 -0.4671413  1.4725803
#> 4 -0.56524992 -1.7978994 -0.09729239 -0.3406456  1.5086451  0.9476612
#> 5 -1.76287295  1.5194596 -1.70743824  0.5456251  0.5943481 -0.5587214
#> 6  0.36938974 -0.1527224 -1.68695056  0.5816300 -0.2217180 -0.3753555
#>          B01         B02
#> 1  1.2255200  0.23789421
#> 2  0.9705887  1.10304905
#> 3 -0.6258206  0.22486129
#> 4  0.6204181  0.83084207
#> 5 -1.1227747 -0.09998736
#> 6  0.3387417 -1.02361703