Convert a DataFrame to a Multi-Frame Matrix

Converts a DataFrame to a matrix where each row is a single time point, and each column is a single well. The wells are in index order, indexing across the rows so the first values are all from row A.

well_df_to_mat_frames(data, values_col, time_col, well_col = "well")

Arguments

data

A DataFrame to be converted.

values_col

The name of the column containing the values.

time_col

The name of the column containing the time values.

well_col

The name of the column with the well IDs.

Value

a matrix

Examples

df_list <- lapply(1:10, function(x) {
  df <- wellr::well_plate()
  df$value <- rnorm(96)
  df$frame <- x
  df
})

df_frames <- do.call(rbind, df_list)

# convert the data frame to multi-frame matrix
mat <- well_df_to_mat_frames(
  data = df_frames,
  values_col = "value",
  time_col = "frame"
)

head(mat[, 1:14])
#>          A01          A02        A03        A04        A05        A06
#> 1 1.01581551  0.798678369 -0.7429861 -0.5837007 -0.0566757  1.5837342
#> 2 0.32134291 -0.424143782 -0.7197842 -1.5268122  0.6108614  0.6597055
#> 3 0.64613620 -0.084433172 -0.1158479 -0.1791726 -1.1572054 -0.2203576
#> 4 0.49341494 -0.655991148  3.1581931 -1.4023169 -0.5122488 -1.3507710
#> 5 0.64507353  0.703907677 -1.5040521 -0.2087806  0.0324386 -1.9530907
#> 6 0.02861699  0.006214292  0.2980690 -0.3759319 -0.8591051 -0.6564904
#>          A07        A08        A09         A10         A11        A12
#> 1 0.73329131  0.3583512 -0.4115058  0.05544703  1.07734073 -0.4012610
#> 2 0.71398372  0.2504576 -0.0939071 -2.08693265  0.66457340 -0.0982727
#> 3 0.91687509  1.3607545  1.6670208  0.18247147  0.34069354 -1.3108776
#> 4 0.17611019 -0.4044742 -0.8265562  0.81319180 -0.05829005  2.2680224
#> 5 0.09116207 -0.1850269  0.2094597  1.81972568  1.27001423 -0.1842145
#> 6 0.14367220  0.7655407  1.4023732 -0.21279338  0.02147185  0.3651401
#>          B01        B02
#> 1 -0.4207703  2.0483690
#> 2  1.1197310  1.2550409
#> 3 -1.5744996  2.4607739
#> 4 -0.5378996  0.8273344
#> 5 -1.9114990 -0.2203366
#> 6 -1.7931829  1.3074243