Convert a DataFrame to a Multi-Frame Matrix

Converts a DataFrame to a matrix where each row is a single time point, and each column is a single well. The wells are in index order, indexing across the rows so the first values are all from row A.

well_df_to_mat_frames(data, values_col, time_col, well_col = "well")

Arguments

data

A DataFrame to be converted.

values_col

The name of the column containing the values.

time_col

The name of the column containing the time values.

well_col

The name of the column with the well IDs.

Value

a matrix

Examples

df_list <- lapply(1:10, function(x) {
  df <- wellr::well_plate()
  df$value <- rnorm(96)
  df$frame <- x
  df
})

df_frames <- do.call(rbind, df_list)

# convert the data frame to multi-frame matrix
mat <- well_df_to_mat_frames(
  data = df_frames,
  values_col = "value",
  time_col = "frame"
)

head(mat[, 1:14])
#>          A01         A02        A03         A04        A05         A06
#> 1 -0.0600245  0.98010822  0.2320037  0.01893782  2.3092180 -0.08299054
#> 2 -0.1497478  0.55377560  0.3333586 -2.75658732 -2.0077356  0.71795133
#> 3  0.5911717 -0.05878625  1.2692182 -0.86900961 -1.3947834  1.27014264
#> 4 -0.1418243 -0.91806354  0.7862419 -0.65691116 -1.7595109  1.43732419
#> 5  0.9200295 -2.01160491  0.2445607 -1.83869304 -1.3851457 -0.07439222
#> 6 -1.2410674  1.54531516 -0.3412501  1.14564270 -0.0245495 -2.65825506
#>            A07        A08         A09        A10         A11        A12
#> 1  0.330725647  0.5619056 -0.02030516 -0.8998963  1.09382448  0.1890344
#> 2 -1.542399448 -0.3449844  0.35877133  1.1191548  0.31926579  0.8104780
#> 3  0.974432430  0.3590505 -1.62533147 -0.2048609  0.04530559  0.6646936
#> 4  0.003949794 -0.2872107  0.30144457  0.4513773 -0.70144270 -0.4564204
#> 5 -0.873762406  0.2623955 -0.26816305  0.5059794  0.49753467 -0.9957383
#> 6 -0.398699806 -1.7851058  0.63203443  0.5746814  0.95149336  0.8819137
#>           B01         B02
#> 1 -1.77417580  0.68048280
#> 2 -0.81551765  1.26834603
#> 3  0.81979282 -0.05233303
#> 4 -0.06483601  1.63363430
#> 5 -0.05778348 -0.65449294
#> 6 -1.04994583  0.76569590