Convert a DataFrame to a Multi-Frame Matrix

Converts a DataFrame to a matrix where each row is a single time point, and each column is a single well. The wells are in index order, indexing across the rows so the first values are all from row A.

well_df_to_mat_frames(data, values_col, time_col, well_col = "well")

Arguments

data

A DataFrame to be converted.

values_col

The name of the column containing the values.

time_col

The name of the column containing the time values.

well_col

The name of the column with the well IDs.

Value

a matrix

Examples

df_list <- lapply(1:10, function(x) {
  df <- wellr::well_plate()
  df$value <- rnorm(96)
  df$frame <- x
  df
})

df_frames <- do.call(rbind, df_list)

# convert the data frame to multi-frame matrix
mat <- well_df_to_mat_frames(
  data = df_frames,
  values_col = "value",
  time_col = "frame"
)

head(mat[, 1:14])
#>          A01        A02         A03         A04        A05        A06
#> 1  0.2581878 -2.2861722 -0.55412818  0.64188363 -0.7439908  0.4588740
#> 2 -0.9712720  0.3546679  0.66482482  1.32802390 -0.6499805  0.2667922
#> 3 -1.6078718  2.9774545  0.45598173  1.65109016 -0.3149438  0.9505107
#> 4  0.4753214  2.5333710  0.07958926  0.06863804  0.3815656 -1.2884867
#> 5 -2.2915374  0.3232935  0.18163955  1.01995858  1.5330821 -0.6326779
#> 6  1.2425654 -0.9794863 -0.24335174 -0.85428061  0.3109764  0.4392437
#>           A07         A08        A09        A10        A11        A12
#> 1  1.76536519  0.81592739 -0.1402353 -1.0700655  0.6227480  1.4772357
#> 2  1.27911739  0.02650361 -0.3158165  0.3459148  0.1962642 -1.1226552
#> 3 -0.49869671 -0.17182422  1.0635032  0.5718331  0.6289902 -2.0450238
#> 4 -0.82438450  0.97068598 -0.1240902  0.6619636 -0.4231706  0.4519824
#> 5  0.41923498 -0.24858201  0.4337153 -0.6458063 -0.3931922 -1.5796924
#> 6 -0.01739982 -1.05312860  0.4111450  0.5241284 -0.1396772 -0.1164553
#>           B01         B02
#> 1  1.05614586  0.59606106
#> 2  0.07495203 -1.41323906
#> 3 -0.98534449  0.02290633
#> 4  0.77460430 -0.71577505
#> 5 -0.70085495  0.64174294
#> 6  0.70688488  0.99357319