Convert a DataFrame to a Multi-Frame Matrix

Converts a DataFrame to a matrix where each row is a single time point, and each column is a single well. The wells are in index order, indexing across the rows so the first values are all from row A.

well_df_to_mat_frames(data, values_col, time_col, well_col = "well")

Arguments

data

A DataFrame to be converted.

values_col

The name of the column containing the values.

time_col

The name of the column containing the time values.

well_col

The name of the column with the well IDs.

Value

a matrix

Examples

df_list <- lapply(1:10, function(x) {
  df <- wellr::well_plate()
  df$value <- rnorm(96)
  df$frame <- x
  df
})

df_frames <- do.call(rbind, df_list)

# convert the data frame to multi-frame matrix
mat <- well_df_to_mat_frames(
  data = df_frames,
  values_col = "value",
  time_col = "frame"
)

head(mat[, 1:14])
#>          A01        A02         A03        A04        A05          A06
#> 1  0.2284635 -1.0406994  0.15881841  1.0324903 -1.4316924  0.166457926
#> 2  0.3592382 -0.6050023  0.51867800 -0.4749563  0.8750359  1.190569044
#> 3 -0.6570398 -1.5413927 -0.04226169 -0.8666088  1.0365965 -0.486704898
#> 4 -2.3866757 -0.6636468  0.07143175 -0.3407216  0.6233673 -0.907271792
#> 5  0.3891984  1.3629045  1.54394774  0.6765323  0.2562804  0.003041776
#> 6 -0.3239639  1.6940002  0.78608421  0.4699494  0.5754680  0.001003314
#>          A07        A08        A09         A10        A11         A12
#> 1  0.1760394  0.7387742  0.2383358  0.89971997 -0.7992602  0.24364025
#> 2 -0.2770317 -0.3117500  0.2773229 -0.30273892  0.6712841 -0.05754948
#> 3  0.8549770 -0.1390945 -0.3977474 -0.42050718  2.2638416 -0.54820427
#> 4 -1.3125177 -1.3055815  1.7615396  0.03725285  0.0765215 -0.10000575
#> 5 -1.6944970  0.6943570  0.8262088 -1.64291941  0.0130576  1.12771976
#> 6  0.1797080  1.5396339  1.8955852  0.45908058  1.1081485  0.09351624
#>           B01          B02
#> 1  1.47291823 -0.638890809
#> 2 -1.04637514  1.511631838
#> 3 -1.58527265  0.453509048
#> 4  0.35681471 -0.948048752
#> 5 -0.94838260  0.002719301
#> 6 -0.09126302  0.337147476