Convert a DataFrame to a Multi-Frame Matrix

Converts a DataFrame to a matrix where each row is a single time point, and each column is a single well. The wells are in index order, indexing across the rows so the first values are all from row A.

well_df_to_mat_frames(data, values_col, time_col, well_col = "well")

Arguments

data

A DataFrame to be converted.

values_col

The name of the column containing the values.

time_col

The name of the column containing the time values.

well_col

The name of the column with the well IDs.

Value

a matrix

Examples

df_list <- lapply(1:10, function(x) {
  df <- wellr::well_plate()
  df$value <- rnorm(96)
  df$frame <- x
  df
})

df_frames <- do.call(rbind, df_list)

# convert the data frame to multi-frame matrix
mat <- well_df_to_mat_frames(
  data = df_frames,
  values_col = "value",
  time_col = "frame"
)

head(mat[, 1:14])
#>           A01        A02        A03        A04         A05        A06
#> 1  1.20170990  0.2105587  0.1608061  0.2733101 -0.10223316 -1.4949727
#> 2  1.88833362 -0.1468194 -0.4446920  1.1391203  1.46721909  0.9456883
#> 3 -1.00739202  1.2908298  0.1765323 -0.2587392  0.01984256  0.6939891
#> 4 -0.24708192  0.1258584  0.2950776 -0.4845968  0.20526422 -1.2390409
#> 5  1.88975668 -0.7164362 -1.7707019 -1.1443060  0.08810185 -1.0476950
#> 6 -0.04531632 -1.3258721  2.1347665  0.1384850  0.86356928 -1.1807443
#>          A07         A08         A09        A10        A11         A12
#> 1  1.3916555  0.04575799 -0.15413220 -0.7971900  1.9436032  0.66965527
#> 2 -0.4962779 -0.27462920  0.76066551 -0.4136831  0.3302123  0.60523285
#> 3 -0.2357141 -1.17457301  0.02589879 -0.3973265  0.7573535  0.06665123
#> 4  0.0630068 -0.79251978 -0.70850979  0.6021083  2.2310142 -1.49784417
#> 5 -0.1728812  1.54730828 -1.26774656  0.9458598 -1.0237523 -0.67067402
#> 6 -0.7970798 -0.55232434 -0.65345978 -0.1143681  1.1259959  0.55596188
#>           B01        B02
#> 1  0.65416819  0.8894311
#> 2  0.07960537 -0.4653487
#> 3  0.24846211 -0.4821990
#> 4 -0.32745772  0.5137077
#> 5  0.34474690  1.4396535
#> 6 -0.11465481 -0.8451918