Convert a DataFrame to a Multi-Frame Matrix

Converts a DataFrame to a matrix where each row is a single time point, and each column is a single well. The wells are in index order, indexing across the rows so the first values are all from row A.

well_df_to_mat_frames(data, values_col, time_col, well_col = "well")

Arguments

data

A DataFrame to be converted.

values_col

The name of the column containing the values.

time_col

The name of the column containing the time values.

well_col

The name of the column with the well IDs.

Value

a matrix

Examples

df_list <- lapply(1:10, function(x) {
  df <- wellr::well_plate()
  df$value <- rnorm(96)
  df$frame <- x
  df
})

df_frames <- do.call(rbind, df_list)

# convert the data frame to multi-frame matrix
mat <- well_df_to_mat_frames(
  data = df_frames,
  values_col = "value",
  time_col = "frame"
)

head(mat[, 1:14])
#>          A01        A02        A03        A04         A05         A06
#> 1  0.7063928 -0.4825054 -0.1851043  1.2851269  1.00721895  0.09787305
#> 2 -0.1644400  0.2811784  1.0212480 -1.0158528  2.06759733  0.39312949
#> 3  0.4624407  0.5568407  0.2158756  0.5434302 -1.21607838  0.94971218
#> 4  0.4778484  0.1842325 -0.1276888  1.5347031  0.67369117  1.13472739
#> 5 -0.1002050  0.2636021  2.1149058 -0.2966024 -0.30574869 -1.34552681
#> 6 -0.1471720 -0.4414612  0.4050315  0.6924452  0.01180492  2.10088725
#>          A07         A08        A09        A10         A11        A12
#> 1 -0.4154645 -1.27250548 -1.3221680  0.5595087  0.06633607  0.1853254
#> 2  1.0790885  0.90891833 -0.4911071  0.4218090 -0.92159241 -0.3451609
#> 3 -0.1116874  0.09658205  1.3891614 -1.1063397  1.11100598  2.6361642
#> 4 -0.3169926 -0.58274685  0.3660686 -1.4787741  0.28837138 -0.3024066
#> 5 -0.4358754 -0.29680052  2.2407034  0.4162219  0.24553684 -0.3509597
#> 6  0.5177658 -0.99194569  0.2277136  0.3060016 -0.50417266  1.1146245
#>          B01         B02
#> 1  1.5755156 -0.21795366
#> 2  0.1563832  0.68444033
#> 3  1.8793004  1.31549398
#> 4 -0.6878187  0.14673382
#> 5 -0.1656051  0.02588487
#> 6  1.2333764  0.73111316