Turn a Multi-Frame Matrix into a DataFrame

Takes a matrix where each row is a time point and each column is a well and converts it to a DataFrame containing a column for the well, a column for the row, a column for the column, a column for the time and a column for the values.

well_mat_frames_to_df(mat, value = "value")

Arguments

mat

A matrix where each column is a well and each row is a frame from a multi time point experiment.

value

Name for the column containing the values in the resulting data frame.

Value

a tibble

Examples

df_list <- lapply(1:10, function(x) {
  df <- wellr::well_plate()
  df$value <- rnorm(96)
  df$frame <- x
  df
})

df_frames <- do.call(rbind, df_list)

# convert the data frame to multi-frame matrix
mat <- well_df_to_mat_frames(
  data = df_frames,
  value = value,
  frame = frame,
  well = well
)

head(mat[, 1:14])
#>            A01        A02        A03          A04        A05         A06
#> 1 -0.544461136  0.1619787  0.4152037  0.463773595 -1.0962935 -0.28623970
#> 2 -0.463011911 -0.3115579 -2.1542090  0.134327521  1.1811777 -0.07273752
#> 3 -0.271757985  0.5520918  1.4471403 -1.537208875 -0.4031708 -0.13072013
#> 4  0.006076236  1.1558658  0.7429154  0.415734090 -0.4978477  0.64955912
#> 5  0.041019444  0.1703206  0.2415014  0.428092637  0.9319746  0.91197125
#> 6  0.764214970  0.8563346  2.2933351 -0.002874673 -1.1958111 -0.68617932
#>          A07        A08        A09        A10         A11         A12
#> 1  1.9390343 -0.6358912 -0.4909112  0.3050033 -0.00947116  0.04585014
#> 2 -0.2344913 -0.0679682 -0.7284801  1.2567631 -1.14333791 -0.51850923
#> 3  1.3032103  1.0303677  0.2192496  0.8355968 -0.34753454  0.83453663
#> 4  0.7084372 -0.1622757  1.0512204 -0.7125816 -0.60457640 -1.86711290
#> 5  0.2739177 -0.6460326 -0.8461900  0.2987841  0.23948024  0.22982555
#> 6  0.5575032  0.8194711  0.1266798  0.6920469 -1.95182047  0.30983115
#>          B01         B02
#> 1 -0.7581075 -0.80860270
#> 2 -0.8342991 -0.85104607
#> 3 -0.5798804 -0.02414068
#> 4  0.9475615 -1.91724811
#> 5  0.8172593 -0.39530479
#> 6  0.7472384  1.26659721

# convert the matrix back to a dataframe
well_mat_frames_to_df(mat)
#> # A tibble: 960 × 5
#>    frame well   value   row   col
#>    <chr> <chr>  <dbl> <dbl> <dbl>
#>  1 1     A01   -0.544     1     1
#>  2 1     A02    0.162     1     2
#>  3 1     A03    0.415     1     3
#>  4 1     A04    0.464     1     4
#>  5 1     A05   -1.10      1     5
#>  6 1     A06   -0.286     1     6
#>  7 1     A07    1.94      1     7
#>  8 1     A08   -0.636     1     8
#>  9 1     A09   -0.491     1     9
#> 10 1     A10    0.305     1    10
#> # … with 950 more rows