Turn a Multi-Frame Matrix into a DataFrame

Takes a matrix where each row is a time point and each column is a well and converts it to a DataFrame containing a column for the well, a column for the row, a column for the column, a column for the time and a column for the values.

well_mat_frames_to_df(mat, value = "value")

Arguments

mat

A matrix where each column is a well and each row is a frame from a multi time point experiment.

value

Name for the column containing the values in the resulting data frame.

Value

a tibble

Examples

df_list <- lapply(1:10, function(x) {
  df <- wellr::well_plate()
  df$value <- rnorm(96)
  df$frame <- x
  df
})

df_frames <- do.call(rbind, df_list)

# convert the data frame to multi-frame matrix
mat <- well_df_to_mat_frames(
  data = df_frames,
  value = value,
  frame = frame,
  well = well
)

head(mat[, 1:14])
#>          A01         A02        A03        A04         A05        A06
#> 1 -0.4738032  1.35144770  0.8892552  1.2643973 -0.73396242 -0.9510534
#> 2  2.3946031  1.71761939 -0.4360176  0.4780144  2.99687104  2.7197489
#> 3  0.1448724 -0.01278213  0.5044266 -0.4144351 -0.35167900 -0.6143909
#> 4 -0.5349269  0.98478974  0.1976185 -1.4893496  0.09862518  0.3716712
#> 5  0.1384005  1.79959990 -0.3080236  0.8540532 -1.06193483  0.9732670
#> 6  0.1648944 -0.39020837  1.1851747  0.3437548 -0.66609248  1.3543667
#>           A07         A08        A09         A10        A11        A12
#> 1  0.20281673 -0.07427421 -0.3218322  0.04986166  0.8874111  0.9659932
#> 2  1.34703081 -0.25000596 -0.2789840 -1.49938009  0.2672565 -1.3387104
#> 3  0.02569057 -0.99674993 -0.5016040 -0.35195687 -0.7243735 -0.1475845
#> 4  0.37640890  1.67703239 -0.6102133  0.29415926 -0.2371453 -0.2478960
#> 5  0.61093612  0.41895369 -0.3920256  1.14835697 -0.3325904  1.3565718
#> 6 -0.05307678 -0.13922586 -0.5908346 -1.09905420  1.2098906  1.2719801
#>          B01         B02
#> 1  0.5790681 -0.47130368
#> 2 -0.1242292 -0.15957617
#> 3  1.7002436 -0.08630022
#> 4  0.3438374 -0.49181951
#> 5  0.5860337 -0.75881610
#> 6  2.5546772  0.28712269

# convert the matrix back to a dataframe
well_mat_frames_to_df(mat)
#> # A tibble: 960 × 5
#>    frame well    value   row   col
#>    <chr> <chr>   <dbl> <dbl> <dbl>
#>  1 1     A01   -0.474      1     1
#>  2 1     A02    1.35       1     2
#>  3 1     A03    0.889      1     3
#>  4 1     A04    1.26       1     4
#>  5 1     A05   -0.734      1     5
#>  6 1     A06   -0.951      1     6
#>  7 1     A07    0.203      1     7
#>  8 1     A08   -0.0743     1     8
#>  9 1     A09   -0.322      1     9
#> 10 1     A10    0.0499     1    10
#> # … with 950 more rows
#> # ℹ Use `print(n = ...)` to see more rows