Turn a Multi-Frame Matrix into a DataFrame

Takes a matrix where each row is a time point and each column is a well and converts it to a DataFrame containing a column for the well, a column for the row, a column for the column, a column for the time and a column for the values.

well_mat_frames_to_df(mat, value = "value")

Arguments

mat

A matrix where each column is a well and each row is a frame from a multi time point experiment.

value

Name for the column containing the values in the resulting data frame.

Value

a tibble

Examples

df_list <- lapply(1:10, function(x) {
  df <- wellr::well_plate()
  df$value <- rnorm(96)
  df$frame <- x
  df
})

df_frames <- do.call(rbind, df_list)

# convert the data frame to multi-frame matrix
mat <- well_df_to_mat_frames(
  data = df_frames,
  value = value,
  frame = frame,
  well = well
)

head(mat[, 1:14])
#>           A01        A02         A03         A04         A05        A06
#> 1 -0.63557112  0.6916144  1.07965917  0.05972293 -1.53131422 -0.6856178
#> 2 -0.00727633  0.7433961 -0.67684555 -0.61151832  0.45656551 -0.2411697
#> 3  1.15466286 -0.2729299  1.45229268  0.45280110  0.70955712 -2.3256108
#> 4  1.36093225 -1.2797319  0.01180026 -1.25042398  0.07681505  1.8008484
#> 5 -1.94139042  0.1097932 -0.20569263 -1.17188137  0.99649318 -1.7981185
#> 6  1.10080494  1.8819148  0.15562358 -0.00567360  2.44568191 -0.1244198
#>          A07         A08        A09        A10        A11        A12        B01
#> 1 -0.5952482  0.83623062  0.1623871 -0.1835196  0.4927572  0.2170598  0.3129632
#> 2  0.6231401  0.71925479  1.4350043  0.6556423 -1.6352996 -0.8970386  0.2203321
#> 3 -1.7335697 -0.45659942 -1.1426727  0.8350123  0.6376704  1.1853770  0.5666179
#> 4 -0.9795662 -0.06192444 -1.5152435  1.4768724 -0.3665658 -0.8174518 -1.1978645
#> 5  0.9797437  0.64819869 -0.2034748 -0.1941834 -0.7105851 -1.6388217  1.3853038
#> 6 -0.1976986  1.12645425  0.9579604 -0.4658871 -2.7503618 -0.6478153 -0.8795143
#>          B02
#> 1  0.7497549
#> 2 -0.6373224
#> 3 -1.2330757
#> 4 -0.5312240
#> 5 -0.1743104
#> 6  1.3616274

# convert the matrix back to a dataframe
well_mat_frames_to_df(mat)
#> # A tibble: 960 × 5
#>    frame well    value   row   col
#>    <chr> <chr>   <dbl> <dbl> <dbl>
#>  1 1     A01   -0.636      1     1
#>  2 1     A02    0.692      1     2
#>  3 1     A03    1.08       1     3
#>  4 1     A04    0.0597     1     4
#>  5 1     A05   -1.53       1     5
#>  6 1     A06   -0.686      1     6
#>  7 1     A07   -0.595      1     7
#>  8 1     A08    0.836      1     8
#>  9 1     A09    0.162      1     9
#> 10 1     A10   -0.184      1    10
#> # … with 950 more rows