Turn a Multi-Frame Matrix into a DataFrame

Takes a matrix where each row is a time point and each column is a well and converts it to a DataFrame containing a column for the well, a column for the row, a column for the column, a column for the time and a column for the values.

well_mat_frames_to_df(matrix, value_col = "value")

Arguments

matrix

multi-frame matrix to be converted.

value_col

Name to be assigned to the values column.

Value

a tibble

Examples

df_list <- lapply(1:10, function(x) {
  df <- wellr::well_plate()
  df$value <- rnorm(96)
  df$frame <- x
  df
})

df_frames <- do.call(rbind, df_list)

# convert the data frame to multi-frame matrix
mat <- well_df_to_mat_frames(
  data = df_frames,
  values_col = "value",
  time_col = "frame"
)

head(mat[, 1:14])
#>          A01        A02        A03        A04        A05        A06        A07
#> 1 -1.6654706  0.7901578  0.7561179 -0.2616220 -0.3507408  0.6289831  0.6812790
#> 2  1.2743208 -1.1979585 -0.6076561  1.2337663 -0.1625838 -0.2445985 -0.6896880
#> 3  1.1704900 -0.6745243 -1.1897068  2.4749972 -0.1304162  0.4573935  0.7410868
#> 4 -0.8683275 -0.6765670  0.8189652 -0.8709581 -0.9775462  1.1125885  0.3809170
#> 5 -0.6647119  0.6827736 -0.6692808  1.4900652 -0.5237991 -1.0675189  1.0188284
#> 6 -2.4379712  0.1479193 -1.4048017  0.1781716 -0.7092015 -0.1422051  1.5239387
#>          A08        A09        A10        A11        A12        B01        B02
#> 1 -0.4675426  0.0897392  0.5599357  0.5040807 0.61849007 -0.2699518 -1.4037649
#> 2 -1.9586227  1.6941971  0.2797902  1.0799773 0.41277556  0.4284685 -0.4789551
#> 3 -1.1377892 -0.1127659  0.8448476  0.1745116 0.08400286  1.0578116  2.0061674
#> 4  0.3478952  0.2127578  1.2142898 -0.1595507 0.92269376  0.1643007  2.3679058
#> 5 -0.2703969 -0.5057861 -0.7269962 -1.4175281 0.57732739  0.4500914 -1.1587242
#> 6 -1.5017971  0.6728613 -1.0835290  1.1824365 1.20279117  2.1081852  0.3469249

# convert the matrix back to a dataframe
well_mat_frames_to_df(mat)
#> # A tibble: 960 × 5
#>    well    row   col   value frames
#>    <chr> <dbl> <dbl>   <dbl> <chr> 
#>  1 A01       1     1 -1.67   1     
#>  2 A02       1     2  0.790  1     
#>  3 A03       1     3  0.756  1     
#>  4 A04       1     4 -0.262  1     
#>  5 A05       1     5 -0.351  1     
#>  6 A06       1     6  0.629  1     
#>  7 A07       1     7  0.681  1     
#>  8 A08       1     8 -0.468  1     
#>  9 A09       1     9  0.0897 1     
#> 10 A10       1    10  0.560  1     
#> # … with 950 more rows