Turn a Multi-Frame Matrix into a DataFrame

Takes a matrix where each row is a time point and each column is a well and converts it to a DataFrame containing a column for the well, a column for the row, a column for the column, a column for the time and a column for the values.

well_mat_frames_to_df(matrix, value_col = "value")

Arguments

matrix

multi-frame matrix to be converted.

value_col

Name to be assigned to the values column.

Value

a tibble

Examples

df_list <- lapply(1:10, function(x) {
  df <- wellr::well_plate()
  df$value <- rnorm(96)
  df$frame <- x
  df
})

df_frames <- do.call(rbind, df_list)

# convert the data frame to multi-frame matrix
mat <- well_df_to_mat_frames(
  data = df_frames,
  values_col = "value",
  time_col = "frame"
)

head(mat[, 1:14])
#>          A01        A02        A03         A04         A05         A06
#> 1 -0.8734233  0.8637674  0.4444266 -0.04528563 -0.99334847 -1.02918349
#> 2 -2.0421843 -1.0139790  0.4636998  0.42137469 -0.88959404 -1.59450662
#> 3  0.1312754 -0.8862460  0.3743348 -2.79980966  0.01041329  0.68164254
#> 4  0.8884911  0.4661584 -0.4662986 -1.12477739  0.91995945  0.71583031
#> 5  0.5061750  0.4499659 -0.4217256 -0.42594288  0.92811621  0.01667958
#> 6 -1.0771912 -1.7515143 -0.1673767  0.53814609 -0.11982744  1.71266599
#>           A07        A08          A09        A10         A11         A12
#> 1 -0.37443904  0.5839563 -0.006840849 -0.1916447  0.16209680  0.45965901
#> 2  0.03074704 -1.2629977  1.274998004  0.1877768 -0.21857224  0.14288297
#> 3  1.66647299  0.1126792 -1.887921798 -0.4181861  0.09086226  0.47442883
#> 4 -0.40519801  0.5903786 -1.183002597 -0.3392988  1.27205032 -1.28677248
#> 5  0.37208166 -0.3479147  0.010116369 -0.3775378  0.20382581 -0.05665524
#> 6 -0.01401544 -0.8969318 -0.804143440  0.4245010  0.93810612 -0.22351190
#>          B01          B02
#> 1 -0.9402096  0.790685973
#> 2 -0.1434756 -1.798722445
#> 3 -0.0262745 -0.000356382
#> 4  1.9394930  0.601937126
#> 5  0.1564236 -0.755319944
#> 6  0.5674550  0.404429759

# convert the matrix back to a dataframe
well_mat_frames_to_df(mat)
#> # A tibble: 960 × 5
#>    well    row   col    value frames
#>    <chr> <dbl> <dbl>    <dbl> <chr> 
#>  1 A01       1     1 -0.873   1     
#>  2 A02       1     2  0.864   1     
#>  3 A03       1     3  0.444   1     
#>  4 A04       1     4 -0.0453  1     
#>  5 A05       1     5 -0.993   1     
#>  6 A06       1     6 -1.03    1     
#>  7 A07       1     7 -0.374   1     
#>  8 A08       1     8  0.584   1     
#>  9 A09       1     9 -0.00684 1     
#> 10 A10       1    10 -0.192   1     
#> # … with 950 more rows