Turn a Multi-Frame Matrix into a DataFrame

Takes a matrix where each row is a time point and each column is a well and converts it to a DataFrame containing a column for the well, a column for the row, a column for the column, a column for the time and a column for the values.

well_mat_frames_to_df(matrix, value_col = "value")

Arguments

matrix

multi-frame matrix to be converted.

value_col

Name to be assigned to the values column.

Value

a tibble

Examples

df_list <- lapply(1:10, function(x) {
  df <- wellr::well_plate()
  df$value <- rnorm(96)
  df$frame <- x
  df
})

df_frames <- do.call(rbind, df_list)

# convert the data frame to multi-frame matrix
mat <- well_df_to_mat_frames(
  data = df_frames,
  values_col = "value",
  time_col = "frame"
)

head(mat[, 1:14])
#>          A01        A02        A03        A04         A05        A06        A07
#> 1 -0.7482442 -0.7636459  1.2006862  0.6470573 -0.57300446  0.7046011 -1.5352207
#> 2 -0.4833000  0.5758041 -0.5113519 -0.6788055  0.03205957  0.4245795 -1.2247696
#> 3 -0.2722920  0.2606405 -1.4468528  0.5609926  0.90394439 -0.9675409 -0.6997304
#> 4 -0.4141045 -0.4409243  0.6267884 -0.2029656  0.35417862  0.5172140 -0.9035580
#> 5  0.2902798  1.6013542 -0.4709897 -0.5383923  0.05975534 -0.3394548 -0.3977242
#> 6  0.4544109 -0.8232213 -0.5267644  1.2195335  0.88194972 -0.9731060 -1.6686116
#>           A08        A09        A10        A11        A12        B01
#> 1 -0.04325208  0.1406346 -0.8941539 -1.5776179  1.3264130  0.5959883
#> 2 -0.31786105  0.1105535  0.4728512 -1.2487514 -1.3187042  1.1956922
#> 3 -0.88454563 -0.3776897 -0.5463439 -0.9901614 -0.2540261  0.8012406
#> 4 -0.98059803  0.7899416 -1.3809383 -0.5421499  1.7747137 -0.5738893
#> 5  0.47280004 -1.1010412  1.1923654 -1.2180961  2.5588782  0.5692153
#> 6  0.12003096  0.8109721 -1.4500556 -1.1131010  0.5961146 -0.7821940
#>           B02
#> 1  0.09191632
#> 2  1.24111461
#> 3 -0.25374045
#> 4 -1.19485949
#> 5  0.80328312
#> 6 -0.74565919

# convert the matrix back to a dataframe
well_mat_frames_to_df(mat)
#> # A tibble: 960 × 5
#>    well    row   col   value frames
#>    <chr> <dbl> <dbl>   <dbl> <chr> 
#>  1 A01       1     1 -0.748  1     
#>  2 A02       1     2 -0.764  1     
#>  3 A03       1     3  1.20   1     
#>  4 A04       1     4  0.647  1     
#>  5 A05       1     5 -0.573  1     
#>  6 A06       1     6  0.705  1     
#>  7 A07       1     7 -1.54   1     
#>  8 A08       1     8 -0.0433 1     
#>  9 A09       1     9  0.141  1     
#> 10 A10       1    10 -0.894  1     
#> # … with 950 more rows