Turn a Multi-Frame Matrix into a DataFrame

Takes a matrix where each row is a time point and each column is a well and converts it to a DataFrame containing a column for the well, a column for the row, a column for the column, a column for the time and a column for the values.

well_mat_frames_to_df(mat, value = "value")

Arguments

mat

A matrix where each column is a well and each row is a frame from a multi time point experiment.

value

Name for the column containing the values in the resulting data frame.

Value

a tibble

Examples

df_list <- lapply(1:10, function(x) {
  df <- wellr::well_plate()
  df$value <- rnorm(96)
  df$frame <- x
  df
})

df_frames <- do.call(rbind, df_list)

# convert the data frame to multi-frame matrix
mat <- well_df_to_mat_frames(
  data = df_frames,
  value = value,
  frame = frame,
  well = well
)

head(mat[, 1:14])
#>           A01        A02        A03         A04         A05         A06
#> 1 -0.20953042  0.1857957 -0.3890444  0.97650599 -1.06937715  0.05137784
#> 2 -0.67794644  0.7243322 -0.5414146  0.29388795  0.01496669  0.19557679
#> 3  0.43702682 -0.9317031 -0.2970275  0.07205331 -0.55984155  1.44834216
#> 4  2.30350423  1.1382188  0.4146961  1.32164477 -1.02995437  0.48139105
#> 5  0.03128265  0.4120993  1.1149742 -0.30071577  0.06674220  0.92797334
#> 6  1.12082524 -0.5259654  0.8125705  0.72485169 -0.49976672 -0.25048732
#>          A07         A08         A09         A10        A11        A12
#> 1 -0.8814653 -0.07684577 -0.15669533 -0.85803931 -0.4504763  0.5835101
#> 2  0.7836590 -0.53474400 -0.16465419  0.07898742  1.0707826  1.3588487
#> 3  0.2123369  1.53280138  1.70814614  1.57871947  0.6318129 -0.5214480
#> 4 -0.8550662 -0.65438671  0.39296499 -0.44162458  0.6853357  0.3561438
#> 5  0.8884437 -0.67238827 -1.30971478 -0.24362792 -0.5932980 -0.7980778
#> 6  0.5615801 -1.79422570  0.05977768  1.58400278 -1.6672854 -0.4915784
#>          B01        B02
#> 1  0.1981274  2.4630095
#> 2 -1.6611408 -0.8248395
#> 3 -0.6208130 -0.1611163
#> 4 -1.4293484  0.4789799
#> 5 -1.4306612  1.4687295
#> 6 -0.2204067 -0.4042966

# convert the matrix back to a dataframe
well_mat_frames_to_df(mat)
#> # A tibble: 960 × 5
#>    frame well    value   row   col
#>    <chr> <chr>   <dbl> <dbl> <dbl>
#>  1 1     A01   -0.210      1     1
#>  2 1     A02    0.186      1     2
#>  3 1     A03   -0.389      1     3
#>  4 1     A04    0.977      1     4
#>  5 1     A05   -1.07       1     5
#>  6 1     A06    0.0514     1     6
#>  7 1     A07   -0.881      1     7
#>  8 1     A08   -0.0768     1     8
#>  9 1     A09   -0.157      1     9
#> 10 1     A10   -0.858      1    10
#> # … with 950 more rows