Turn a Multi-Frame Matrix into a DataFrame

Takes a matrix where each row is a time point and each column is a well and converts it to a DataFrame containing a column for the well, a column for the row, a column for the column, a column for the time and a column for the values.

well_mat_frames_to_df(mat, value = "value")

Arguments

mat

A matrix where each column is a well and each row is a frame from a multi time point experiment.

value

Name for the column containing the values in the resulting data frame.

Value

a tibble

Examples

df_list <- lapply(1:10, function(x) {
  df <- wellr::well_plate()
  df$value <- rnorm(96)
  df$frame <- x
  df
})

df_frames <- do.call(rbind, df_list)

# convert the data frame to multi-frame matrix
mat <- well_df_to_mat_frames(
  data = df_frames,
  value = value,
  frame = frame,
  well = well
)

head(mat[, 1:14])
#>           A01        A02        A03         A04        A05        A06
#> 1 -0.06470322 -0.5508377 -0.2945410  1.58972077 -0.3371876 -0.4435216
#> 2  0.47069632 -0.5962405  0.5505964 -0.07579364  0.2030867  0.5908723
#> 3 -0.28601678  0.3672039 -1.9618775 -0.62247625 -0.6270785  0.2619259
#> 4 -2.11095824  0.2805448  0.3146802 -0.62191398  1.4216357  0.7149112
#> 5  0.14182117  1.5359570 -0.8785424 -0.11231516  0.2190101 -1.1315455
#> 6  1.66962410 -2.6965043 -2.5254363 -0.25727015 -0.4211587  2.3807423
#>           A07         A08         A09        A10        A11        A12
#> 1  0.13748192  0.53486298 -1.22902252  1.0515961  0.9459115 -0.3517993
#> 2 -0.09375278 -0.03922036  1.40822389 -0.4576450 -1.0425456 -0.9680770
#> 3 -0.10015263  1.75272828  0.18456191 -0.5127223  0.9017586  0.8218298
#> 4 -0.57866612 -0.56305434 -0.99372009  0.7711706 -0.7192764 -1.6351499
#> 5 -1.21262490 -0.04229206  0.98899277  0.5936203 -0.4113707 -0.4781288
#> 6 -1.46848606 -0.54048332  0.03380459 -1.3738568  0.3813701 -0.5374897
#>           B01         B02
#> 1 -0.71964162 -0.42036770
#> 2 -0.02113756 -0.37853138
#> 3  0.67458952 -0.71342173
#> 4  0.72221940 -0.56084101
#> 5  1.34005529 -0.05975355
#> 6  0.16097281 -0.76773159

# convert the matrix back to a dataframe
well_mat_frames_to_df(mat)
#> # A tibble: 960 × 5
#>    frame well    value   row   col
#>    <chr> <chr>   <dbl> <dbl> <dbl>
#>  1 1     A01   -0.0647     1     1
#>  2 1     A02   -0.551      1     2
#>  3 1     A03   -0.295      1     3
#>  4 1     A04    1.59       1     4
#>  5 1     A05   -0.337      1     5
#>  6 1     A06   -0.444      1     6
#>  7 1     A07    0.137      1     7
#>  8 1     A08    0.535      1     8
#>  9 1     A09   -1.23       1     9
#> 10 1     A10    1.05       1    10
#> # … with 950 more rows