Turn a Multi-Frame Matrix into a DataFrame

Takes a matrix where each row is a time point and each column is a well and converts it to a DataFrame containing a column for the well, a column for the row, a column for the column, a column for the time and a column for the values.

well_mat_frames_to_df(mat, value = "value")

Arguments

mat

A matrix where each column is a well and each row is a frame from a multi time point experiment.

value

Name for the column containing the values in the resulting data frame.

Value

a tibble

Examples

df_list <- lapply(1:10, function(x) {
  df <- wellr::well_plate()
  df$value <- rnorm(96)
  df$frame <- x
  df
})

df_frames <- do.call(rbind, df_list)

# convert the data frame to multi-frame matrix
mat <- well_df_to_mat_frames(
  data = df_frames,
  value = value,
  frame = frame,
  well = well
)

head(mat[, 1:14])
#>          A01        A02        A03         A04         A05        A06
#> 1  1.5460007 -1.5064203  0.9438216  1.72225548  0.27252783 -0.6128596
#> 2 -0.4112688 -1.2520003  0.6505394  1.02543177 -0.13965751  1.7517900
#> 3  0.3908736 -0.3227134 -0.4131015  0.02644746 -0.27099076  1.2186707
#> 4 -0.7578093  0.2822174  0.8190441 -0.60922604  0.29986844 -0.7037875
#> 5 -0.2313679 -1.7516243 -0.1999961 -1.19998369 -0.08034153 -1.2684386
#> 6 -0.3305813  0.2605924 -1.8210352 -0.56814260  0.09765807 -0.6462029
#>          A07         A08        A09         A10        A11         A12
#> 1 -0.3503378 -0.00327616  1.2946606  0.40178892  0.2348851 -1.41571628
#> 2  0.7424017  1.08226388 -1.4108109  0.05457555  0.6614975 -0.70853367
#> 3  0.9187533 -0.71264017 -0.8650198 -0.78063817  0.2322470 -0.09549123
#> 4  0.5128703 -0.55062303 -0.2315638  0.21287087  0.4757616  2.69345561
#> 5  1.8804825 -0.43566205 -1.9016486  0.13427059 -0.3828196  0.60671617
#> 6 -0.1472450 -0.33970776  0.5510034 -1.41805586 -0.6921588 -0.13652377
#>          B01         B02
#> 1  1.8801300  1.42260724
#> 2 -0.3222856 -0.14472213
#> 3  0.4653247  1.73424744
#> 4  2.1167516 -0.42447910
#> 5 -0.3421578 -1.43219333
#> 6 -1.0202336  0.02900791

# convert the matrix back to a dataframe
well_mat_frames_to_df(mat)
#> # A tibble: 960 × 5
#>    frame well     value   row   col
#>    <chr> <chr>    <dbl> <dbl> <dbl>
#>  1 1     A01    1.55        1     1
#>  2 1     A02   -1.51        1     2
#>  3 1     A03    0.944       1     3
#>  4 1     A04    1.72        1     4
#>  5 1     A05    0.273       1     5
#>  6 1     A06   -0.613       1     6
#>  7 1     A07   -0.350       1     7
#>  8 1     A08   -0.00328     1     8
#>  9 1     A09    1.29        1     9
#> 10 1     A10    0.402       1    10
#> # … with 950 more rows