Turn a Multi-Frame Matrix into a DataFrame

Takes a matrix where each row is a time point and each column is a well and converts it to a DataFrame containing a column for the well, a column for the row, a column for the column, a column for the time and a column for the values.

well_mat_frames_to_df(mat, value = "value")

Arguments

mat

A matrix where each column is a well and each row is a frame from a multi time point experiment.

value

Name for the column containing the values in the resulting data frame.

Value

a tibble

Examples

df_list <- lapply(1:10, function(x) {
  df <- wellr::well_plate()
  df$value <- rnorm(96)
  df$frame <- x
  df
})

df_frames <- do.call(rbind, df_list)

# convert the data frame to multi-frame matrix
mat <- well_df_to_mat_frames(
  data = df_frames,
  value = value,
  frame = frame,
  well = well
)

head(mat[, 1:14])
#>          A01        A02        A03         A04        A05        A06
#> 1  1.5463993 -0.2421962 -1.6123693 -0.35314778  0.8125947 -0.9437194
#> 2 -0.5157617 -0.5279622 -1.7044154  1.09643940  1.2841207 -2.1155239
#> 3  0.0491758  0.8427578 -1.0032142 -0.02861121  2.0110936 -1.7989878
#> 4 -0.7603104 -0.4420060  0.3923001 -1.73926774  1.3224125 -0.1189050
#> 5 -0.4861222 -0.0208237 -0.1880532 -0.23349418 -0.5874402 -0.9058531
#> 6 -0.9353330  0.2925033 -2.4052081  1.21731880  0.9959307 -0.0946538
#>           A07         A08        A09        A10        A11        A12
#> 1 -2.19739339  0.85767078  0.4943894  0.1940431  0.7239429  0.3546851
#> 2  0.91212080 -0.86486137  1.1740699 -1.8127235  1.5294528 -0.3948431
#> 3 -0.28841971  0.30614963  1.3514423  0.9585073 -0.5356755 -1.1286681
#> 4 -0.94634382  0.45022564 -0.5116930 -0.3852924  0.9797093  2.5000340
#> 5  0.09561014 -0.01916905  0.9984903  0.4427216 -0.6122223 -1.2854940
#> 6 -1.06843970 -0.76279666  0.4500189  0.9408671 -1.8131391  1.4337930
#>          B01        B02
#> 1  2.3462781 -0.6257992
#> 2 -0.5122055 -0.5495654
#> 3 -0.1667893 -0.6889355
#> 4 -0.7209214  0.9731530
#> 5 -0.1762396 -0.3752315
#> 6  1.7550215 -0.7456325

# convert the matrix back to a dataframe
well_mat_frames_to_df(mat)
#> # A tibble: 960 × 5
#>    frame well   value   row   col
#>    <chr> <chr>  <dbl> <dbl> <dbl>
#>  1 1     A01    1.55      1     1
#>  2 1     A02   -0.242     1     2
#>  3 1     A03   -1.61      1     3
#>  4 1     A04   -0.353     1     4
#>  5 1     A05    0.813     1     5
#>  6 1     A06   -0.944     1     6
#>  7 1     A07   -2.20      1     7
#>  8 1     A08    0.858     1     8
#>  9 1     A09    0.494     1     9
#> 10 1     A10    0.194     1    10
#> # … with 950 more rows