Turn a Multi-Frame Matrix into a DataFrame

Takes a matrix where each row is a time point and each column is a well and converts it to a DataFrame containing a column for the well, a column for the row, a column for the column, a column for the time and a column for the values.

well_mat_frames_to_df(matrix, value_col = "value")

Arguments

matrix

multi-frame matrix to be converted.

value_col

Name to be assigned to the values column.

Value

a tibble

Examples

df_list <- lapply(1:10, function(x) {
  df <- wellr::well_plate()
  df$value <- rnorm(96)
  df$frame <- x
  df
})

df_frames <- do.call(rbind, df_list)

# convert the data frame to multi-frame matrix
mat <- well_df_to_mat_frames(
  data = df_frames,
  values_col = "value",
  time_col = "frame"
)
#> Error in well_df_to_mat_frames(data = df_frames, values_col = "value",     time_col = "frame"): could not find function "well_df_to_mat_frames"

head(mat[, 1:14])
#> Error in head(mat[, 1:14]): object 'mat' not found

# convert the matrix back to a dataframe
well_mat_frames_to_df(mat)
#> Error in well_mat_frames_to_df(mat): could not find function "well_mat_frames_to_df"