Turn a Multi-Frame Matrix into a DataFrame

Takes a matrix where each row is a time point and each column is a well and converts it to a DataFrame containing a column for the well, a column for the row, a column for the column, a column for the time and a column for the values.

well_mat_frames_to_df(matrix, value_col = "value")

Arguments

matrix

multi-frame matrix to be converted.

value_col

Name to be assigned to the values column.

Value

a tibble

Examples

df_list <- lapply(1:10, function(x) {
  df <- wellr::well_plate()
  df$value <- rnorm(96)
  df$frame <- x
  df
})

df_frames <- do.call(rbind, df_list)

# convert the data frame to multi-frame matrix
mat <- well_df_to_mat_frames(
  data = df_frames,
  values_col = "value",
  time_col = "frame"
)

head(mat[, 1:14])
#>           A01          A02         A03        A04        A05         A06
#> 1  1.02525477  0.365154332 -0.39493461 -0.9726913 -0.1604071 -0.38998063
#> 2  0.68438480  0.323409345 -0.48928406  1.5340153  0.6935123 -1.31469080
#> 3 -0.09436175 -0.324983807  1.49994277  1.3869427  1.0065113 -0.40399589
#> 4  0.13144314 -0.219870366 -2.00402721  0.2677315  0.0310022 -1.67271949
#> 5  0.96847158 -0.004747716 -0.03225539  0.1018452  0.9917947  0.26571725
#> 6  0.53427502 -0.445898624  0.25940160  1.0466972 -2.4776244 -0.05189281
#>            A07        A08        A09          A10        A11       A12
#> 1  1.977491181 -1.1360993  0.2074829  1.555616497  0.1249009 1.8349096
#> 2  0.009280774  0.4116865 -0.1452014 -1.055209521  0.6108620 0.8856640
#> 3  0.630823264 -1.6017256  0.7270045  1.055128604 -0.3629384 0.8661829
#> 4  0.379810325  1.5263013 -1.0615457  2.512219227 -1.1951640 0.1090534
#> 5  1.556553223  1.8118550  0.8805879  1.700605524 -0.1919098 0.4365835
#> 6 -1.387328557 -0.3688707 -0.1591716  0.005762722 -0.4485610 1.0861358
#>           B01         B02
#> 1 -0.63613276 -0.09901044
#> 2 -0.94569426 -0.32922323
#> 3  1.69389619  0.85924892
#> 4 -1.64684055  0.32697019
#> 5  0.02150693  0.65695011
#> 6  1.25645154  0.86050652

# convert the matrix back to a dataframe
well_mat_frames_to_df(mat)
#> # A tibble: 960 × 5
#>    well    row   col  value frames
#>    <chr> <dbl> <dbl>  <dbl> <chr> 
#>  1 A01       1     1  1.03  1     
#>  2 A02       1     2  0.207 1     
#>  3 A03       1     3 -1.25  1     
#>  4 A04       1     4  0.504 1     
#>  5 A05       1     5  0.133 1     
#>  6 A06       1     6 -0.834 1     
#>  7 A07       1     7 -0.421 1     
#>  8 A08       1     8 -0.939 1     
#>  9 A09       1     9  1.86  1     
#> 10 A10       1    10 -0.172 1     
#> # … with 950 more rows
#> # ℹ Use `print(n = ...)` to see more rows