Turn a Multi-Frame Matrix into a DataFrame

Takes a matrix where each row is a time point and each column is a well and converts it to a DataFrame containing a column for the well, a column for the row, a column for the column, a column for the time and a column for the values.

well_mat_frames_to_df(mat, value = "value")

Arguments

mat

A matrix where each column is a well and each row is a frame from a multi time point experiment.

value

Name for the column containing the values in the resulting data frame.

Value

a tibble

Examples

df_list <- lapply(1:10, function(x) {
  df <- wellr::well_plate()
  df$value <- rnorm(96)
  df$frame <- x
  df
})

df_frames <- do.call(rbind, df_list)

# convert the data frame to multi-frame matrix
mat <- well_df_to_mat_frames(
  data = df_frames,
  value = value,
  frame = frame,
  well = well
)

head(mat[, 1:14])
#>          A01        A02         A03        A04          A05         A06
#> 1 -0.4049020  0.5949912 -0.87607717  0.2865305  0.302642000  1.04887988
#> 2 -1.6252714 -1.1608765 -0.63886023  1.7743678 -1.876726719 -0.52841862
#> 3  0.7354096  0.1594283  0.72864893  0.9108205 -0.117936848  0.54038753
#> 4  0.7866717 -1.4493297  0.01980255  1.9836802  0.066009570 -1.00733276
#> 5  0.3116608  0.4824406 -1.06383749  2.1253503 -0.083175204 -1.18324018
#> 6  0.2246518  0.7191468  1.88289817 -0.1793354 -0.002889819  0.01915511
#>          A07        A08         A09        A10        A11         A12
#> 1 -1.3619021  0.2944692  0.97847894  0.2745498  1.1041964 -0.19078349
#> 2 -0.1786167 -1.1312568  0.13935885 -0.8318116 -0.7164755 -0.33264291
#> 3 -1.1645347 -0.5809355  1.08741918  1.1950023  1.2140726 -0.28217066
#> 4 -1.3641271 -0.1958417 -0.02442940  1.5185178  1.4816236 -0.90472929
#> 5 -2.3586520 -0.6016091 -0.07109225 -0.8575239 -0.7239511  2.09738486
#> 6  0.6701781 -0.3769145 -0.03928754 -0.1257345 -0.8568290 -0.03445456
#>          B01         B02
#> 1  0.4253695  0.05581197
#> 2 -1.9828855  1.13674157
#> 3  0.7467873 -1.58033997
#> 4  0.4165805 -0.58275501
#> 5  0.2138733  0.63610323
#> 6  0.2340077  2.08665629

# convert the matrix back to a dataframe
well_mat_frames_to_df(mat)
#> # A tibble: 960 × 5
#>    frame well   value   row   col
#>    <chr> <chr>  <dbl> <dbl> <dbl>
#>  1 1     A01   -0.405     1     1
#>  2 1     A02    0.595     1     2
#>  3 1     A03   -0.876     1     3
#>  4 1     A04    0.287     1     4
#>  5 1     A05    0.303     1     5
#>  6 1     A06    1.05      1     6
#>  7 1     A07   -1.36      1     7
#>  8 1     A08    0.294     1     8
#>  9 1     A09    0.978     1     9
#> 10 1     A10    0.275     1    10
#> # ℹ 950 more rows