Turn a Multi-Frame Matrix into a DataFrame

Takes a matrix where each row is a time point and each column is a well and converts it to a DataFrame containing a column for the well, a column for the row, a column for the column, a column for the time and a column for the values.

well_mat_frames_to_df(matrix, value_col = "value")

Arguments

matrix

multi-frame matrix to be converted.

value_col

Name to be assigned to the values column.

Value

a tibble

Examples

df_list <- lapply(1:10, function(x) {
  df <- wellr::well_plate()
  df$value <- rnorm(96)
  df$frame <- x
  df
})

df_frames <- do.call(rbind, df_list)

# convert the data frame to multi-frame matrix
mat <- well_df_to_mat_frames(
  data = df_frames,
  values_col = "value",
  time_col = "frame"
)

head(mat[, 1:14])
#>           A01         A02         A03         A04          A05        A06
#> 1 -0.02171925 -0.95942585 -0.01995586 -0.04358983 -1.233104926 -1.1193596
#> 2  0.12996445  0.33621180  0.50913389 -0.03443357 -0.003768792  0.7098970
#> 3 -0.37073523 -0.25411706  1.52053635  1.08135909  0.140575347 -0.5418778
#> 4 -0.33123090  0.26309829 -0.16716747  1.10054440  0.878448918  2.7447232
#> 5  0.02067646  0.52757936 -0.42972067 -0.04674414 -0.222301546 -1.5874307
#> 6  1.00370518 -0.05246249  0.18370787  0.53404467  0.072008974 -3.1249259
#>          A07        A08        A09        A10        A11        A12         B01
#> 1 -1.6019314 -0.5940620 -0.9664351  0.1497564  0.7431113 -0.1880168 -1.30291068
#> 2 -1.7972101  0.6028048 -0.6998660 -0.5757991  1.1538015 -0.8852981 -0.27600389
#> 3 -1.3006834  0.9701053 -0.4360356 -0.2239140  0.5343808  0.1443204  0.01841028
#> 4  0.9930171 -1.0934384 -0.5123722 -0.4354144 -0.2896749  1.3149176  0.47639400
#> 5  1.0740047 -0.4856717  0.1886184  1.5103008  0.6598946  1.8827067  1.71544654
#> 6 -0.6724153  2.0132917  1.3013737 -1.4768120 -1.7428168  0.5730598 -0.21334134
#>          B02
#> 1 -0.7291198
#> 2 -0.6795374
#> 3  0.7522922
#> 4 -0.9773371
#> 5  1.2437846
#> 6 -0.5897081

# convert the matrix back to a dataframe
well_mat_frames_to_df(mat)
#> # A tibble: 960 × 5
#>    well    row   col   value frames
#>    <chr> <dbl> <dbl>   <dbl> <chr> 
#>  1 A01       1     1 -0.0217 1     
#>  2 A02       1     2 -0.959  1     
#>  3 A03       1     3 -0.0200 1     
#>  4 A04       1     4 -0.0436 1     
#>  5 A05       1     5 -1.23   1     
#>  6 A06       1     6 -1.12   1     
#>  7 A07       1     7 -1.60   1     
#>  8 A08       1     8 -0.594  1     
#>  9 A09       1     9 -0.966  1     
#> 10 A10       1    10  0.150  1     
#> # … with 950 more rows