Turn a Multi-Frame Matrix into a DataFrame

Takes a matrix where each row is a time point and each column is a well and converts it to a DataFrame containing a column for the well, a column for the row, a column for the column, a column for the time and a column for the values.

well_mat_frames_to_df(mat, value = "value")

Arguments

mat

A matrix where each column is a well and each row is a frame from a multi time point experiment.

value

Name for the column containing the values in the resulting data frame.

Value

a tibble

Examples

df_list <- lapply(1:10, function(x) {
  df <- wellr::well_plate()
  df$value <- rnorm(96)
  df$frame <- x
  df
})

df_frames <- do.call(rbind, df_list)

# convert the data frame to multi-frame matrix
mat <- well_df_to_mat_frames(
  data = df_frames,
  value = value,
  frame = frame,
  well = well
)

head(mat[, 1:14])
#>           A01         A02           A03        A04        A05        A06
#> 1  0.82759542 -0.76238454 -1.585951e-01 -0.1805180  0.4308612 -0.6681626
#> 2  0.54650558 -0.53496919 -3.967944e-01 -0.9715399  0.1809799  1.4037417
#> 3 -0.06639546  1.96665910 -1.382041e+00 -1.1683980  0.3657946  0.2575305
#> 4  0.18762089 -1.21327604  4.704621e-05 -0.1106626 -0.9372056  0.3136077
#> 5 -0.96120073  0.09733372  2.502439e-01  0.8205370 -0.2176935 -1.4263246
#> 6  0.33915575 -1.57365485 -3.153234e-01  0.7610689 -0.8109333 -1.8662669
#>           A07        A08         A09        A10        A11        A12
#> 1  0.01448441  0.4444391  0.07380785 -1.0264689  0.4702163 -0.2692838
#> 2  0.48285038  0.9540427  1.12592756  1.2703255  0.6987065  0.4301134
#> 3  0.89002119  0.9806120  0.45722618 -1.1195211 -0.9051516 -0.5325209
#> 4  1.24184171  1.9392000  0.53263412 -0.3913815 -1.2586038  0.2320517
#> 5 -1.30320359 -0.5560469 -0.71554101 -0.6876142  0.1548810  0.6970649
#> 6  1.01608849 -0.4534647 -0.12470901  0.7633213 -0.4629668 -0.1454664
#>          B01        B02
#> 1 -0.9962628 -1.2145087
#> 2 -0.5141953  0.6668017
#> 3 -0.3573437 -0.5693581
#> 4  0.5740415 -0.3964887
#> 5 -0.4825041  1.5910877
#> 6 -0.2798388 -0.3222354

# convert the matrix back to a dataframe
well_mat_frames_to_df(mat)
#> # A tibble: 960 × 5
#>    frame well    value   row   col
#>    <chr> <chr>   <dbl> <dbl> <dbl>
#>  1 1     A01    0.828      1     1
#>  2 1     A02   -0.762      1     2
#>  3 1     A03   -0.159      1     3
#>  4 1     A04   -0.181      1     4
#>  5 1     A05    0.431      1     5
#>  6 1     A06   -0.668      1     6
#>  7 1     A07    0.0145     1     7
#>  8 1     A08    0.444      1     8
#>  9 1     A09    0.0738     1     9
#> 10 1     A10   -1.03       1    10
#> # ℹ 950 more rows