Turn a Multi-Frame Matrix into a DataFrame

Takes a matrix where each row is a time point and each column is a well and converts it to a DataFrame containing a column for the well, a column for the row, a column for the column, a column for the time and a column for the values.

well_mat_frames_to_df(mat, value = "value")

Arguments

mat

A matrix where each column is a well and each row is a frame from a multi time point experiment.

value

Name for the column containing the values in the resulting data frame.

Value

a tibble

Examples

df_list <- lapply(1:10, function(x) {
  df <- wellr::well_plate()
  df$value <- rnorm(96)
  df$frame <- x
  df
})

df_frames <- do.call(rbind, df_list)

# convert the data frame to multi-frame matrix
mat <- well_df_to_mat_frames(
  data = df_frames,
  value = value,
  frame = frame,
  well = well
)

head(mat[, 1:14])
#>          A01        A02        A03         A04         A05        A06
#> 1  0.8084397 -0.2642820  0.3883343 -0.03513836 -0.08605798 -1.4575771
#> 2  0.3665833  0.4650803  0.2789653  1.10183960  0.43038859  2.4005091
#> 3 -0.2803806  0.3547909 -0.3093496  0.01384254 -0.36852494 -0.8978822
#> 4  0.8444679 -1.2650094 -0.3834648 -1.00245346 -1.05613861 -0.8758397
#> 5  1.0006464  0.1724763 -0.7659607 -0.70911588 -0.05823860  0.9401716
#> 6  0.4552226  0.9268753  0.1589732  0.48262714 -0.89104621  0.1930124
#>          A07         A08        A09         A10         A11        A12
#> 1 -0.1887011 -0.47859865 -1.2738461  0.29033830  0.03195513 -0.4356607
#> 2  1.0994260  0.03736182  0.3800342 -1.91356389 -0.01715112 -0.5580678
#> 3 -0.4040254 -0.26532681 -1.0751001 -0.07895490 -0.42327527  0.5906426
#> 4  0.2731975  0.22830376  0.0440043 -0.04975978 -0.81774789  0.6556455
#> 5  0.2023073 -0.13119818 -0.5149741 -0.57181454  1.67509394  0.4037251
#> 6  1.9339301 -0.82732766  1.0590553 -0.77320967  0.71987093 -0.3597991
#>           B01        B02
#> 1 -1.50784893  0.5246193
#> 2  0.84940572 -0.2180299
#> 3  1.48642887 -0.6310462
#> 4  0.33408147  0.8981800
#> 5  0.75709996  0.2224181
#> 6 -0.08741618 -0.2194066

# convert the matrix back to a dataframe
well_mat_frames_to_df(mat)
#> # A tibble: 960 × 5
#>    frame well    value   row   col
#>    <chr> <chr>   <dbl> <dbl> <dbl>
#>  1 1     A01    0.808      1     1
#>  2 1     A02   -0.264      1     2
#>  3 1     A03    0.388      1     3
#>  4 1     A04   -0.0351     1     4
#>  5 1     A05   -0.0861     1     5
#>  6 1     A06   -1.46       1     6
#>  7 1     A07   -0.189      1     7
#>  8 1     A08   -0.479      1     8
#>  9 1     A09   -1.27       1     9
#> 10 1     A10    0.290      1    10
#> # … with 950 more rows