Turn a Multi-Frame Matrix into a DataFrame

Takes a matrix where each row is a time point and each column is a well and converts it to a DataFrame containing a column for the well, a column for the row, a column for the column, a column for the time and a column for the values.

well_mat_frames_to_df(mat, value = "value")

Arguments

mat

A matrix where each column is a well and each row is a frame from a multi time point experiment.

value

Name for the column containing the values in the resulting data frame.

Value

a tibble

Examples

df_list <- lapply(1:10, function(x) {
  df <- wellr::well_plate()
  df$value <- rnorm(96)
  df$frame <- x
  df
})

df_frames <- do.call(rbind, df_list)

# convert the data frame to multi-frame matrix
mat <- well_df_to_mat_frames(
  data = df_frames,
  value = value,
  frame = frame,
  well = well
)

head(mat[, 1:14])
#>          A01         A02        A03         A04         A05        A06
#> 1 -0.5572600  1.51211242  0.6224009  0.30272123  1.17233767  1.5531962
#> 2  0.8377243  0.12156005 -1.3261915  0.04630191 -0.46796983  0.1915229
#> 3  1.2642020  1.48792761  0.5614011 -0.83044428 -1.65601807 -0.3099298
#> 4 -0.5132819 -0.35325712  0.3664852 -0.11779952  0.33190139 -1.0096405
#> 5  1.6281268 -0.43620822  1.2898311  0.27669696 -0.03591466 -1.2655332
#> 6  0.7218134  0.09804404 -0.4125675 -0.93235342 -0.88539716 -0.7022314
#>          A07         A08        A09         A10         A11        A12
#> 1 -0.4247101  0.04338528  1.3990278 -0.68437607 -1.37360898 -0.1935581
#> 2 -0.3199281 -1.09760117  0.4183348  0.36574396  0.57937107  0.3233945
#> 3  1.1181474  0.86141200 -1.0321356  0.89187198 -0.06152171 -0.4728386
#> 4 -0.1017993  0.56205924  1.1208011  0.95364939 -0.03446054 -0.3990850
#> 5 -0.1609155  0.56266505 -0.9185062  0.01652333 -1.86757133 -0.6143828
#> 6 -0.9352517 -0.35129734  0.1422470 -0.57223912 -0.67322181  0.3749281
#>           B01        B02
#> 1 -0.28150782 -1.2721752
#> 2 -0.93518243  0.3811330
#> 3 -0.31789684  0.2140433
#> 4 -1.90362651  1.1338837
#> 5 -0.06583993 -2.8622949
#> 6 -0.45917936  0.1628385

# convert the matrix back to a dataframe
well_mat_frames_to_df(mat)
#> # A tibble: 960 × 5
#>    frame well    value   row   col
#>    <chr> <chr>   <dbl> <dbl> <dbl>
#>  1 1     A01   -0.557      1     1
#>  2 1     A02    1.51       1     2
#>  3 1     A03    0.622      1     3
#>  4 1     A04    0.303      1     4
#>  5 1     A05    1.17       1     5
#>  6 1     A06    1.55       1     6
#>  7 1     A07   -0.425      1     7
#>  8 1     A08    0.0434     1     8
#>  9 1     A09    1.40       1     9
#> 10 1     A10   -0.684      1    10
#> # … with 950 more rows