Turn a Multi-Frame Matrix into a DataFrame

Takes a matrix where each row is a time point and each column is a well and converts it to a DataFrame containing a column for the well, a column for the row, a column for the column, a column for the time and a column for the values.

well_mat_frames_to_df(mat, value = "value")

Arguments

mat

A matrix where each column is a well and each row is a frame from a multi time point experiment.

value

Name for the column containing the values in the resulting data frame.

Value

a tibble

Examples

df_list <- lapply(1:10, function(x) {
  df <- wellr::well_plate()
  df$value <- rnorm(96)
  df$frame <- x
  df
})

df_frames <- do.call(rbind, df_list)

# convert the data frame to multi-frame matrix
mat <- well_df_to_mat_frames(
  data = df_frames,
  value = value,
  frame = frame,
  well = well
)

head(mat[, 1:14])
#>          A01        A02         A03        A04        A05        A06
#> 1  1.4264812  0.8984647 -0.04106031 -0.1993526  0.5032161 -1.1670985
#> 2  0.6348197 -0.7053817  0.32884599 -1.7829621 -0.3087780 -0.1074808
#> 3 -1.9378709 -2.5848086 -0.27451129 -0.5965032  0.9128250 -1.3036331
#> 4 -1.2403383  1.0917519  0.08965287  0.1089275 -0.1181917  0.4072255
#> 5  0.8406746 -0.7888859 -0.46585912 -1.9032128 -0.9509578 -0.1686726
#> 6  1.3449904  0.9463911  0.90574727  1.0226566 -1.2206684  1.6064518
#>           A07         A08        A09         A10        A11         A12
#> 1 -0.28066796 -0.85568774 -1.2165774 -2.09918667  0.2183106  0.94433025
#> 2 -0.13359692 -0.96460775  1.0673850  1.14602070  1.2212745 -1.38129638
#> 3  0.38536660 -0.64010919 -0.5479514 -0.63999474  0.2288159 -0.30033546
#> 4 -0.61493884 -0.50044646 -0.9313182 -0.86731594 -0.6968665 -0.04455186
#> 5 -0.01318064 -0.53217784 -0.6399406 -0.08645324 -0.8620550  1.02837450
#> 6 -1.62252966  0.07898588 -0.3174258 -0.94054506 -1.0871668 -0.42560523
#>          B01         B02
#> 1 -0.2934455  0.36886388
#> 2 -0.1782003  1.75383951
#> 3  1.1631301  0.05831165
#> 4 -3.3066484 -0.57973964
#> 5 -0.3804453  3.38326750
#> 6  0.3684320  0.15355260

# convert the matrix back to a dataframe
well_mat_frames_to_df(mat)
#> # A tibble: 960 × 5
#>    frame well    value   row   col
#>    <chr> <chr>   <dbl> <dbl> <dbl>
#>  1 1     A01    1.43       1     1
#>  2 1     A02    0.898      1     2
#>  3 1     A03   -0.0411     1     3
#>  4 1     A04   -0.199      1     4
#>  5 1     A05    0.503      1     5
#>  6 1     A06   -1.17       1     6
#>  7 1     A07   -0.281      1     7
#>  8 1     A08   -0.856      1     8
#>  9 1     A09   -1.22       1     9
#> 10 1     A10   -2.10       1    10
#> # … with 950 more rows