Turn a Multi-Frame Matrix into a DataFrame

Takes a matrix where each row is a time point and each column is a well and converts it to a DataFrame containing a column for the well, a column for the row, a column for the column, a column for the time and a column for the values.

well_mat_frames_to_df(matrix, value_col = "value")

Arguments

matrix

multi-frame matrix to be converted.

value_col

Name to be assigned to the values column.

Value

a tibble

Examples

df_list <- lapply(1:10, function(x) {
  df <- wellr::well_plate()
  df$value <- rnorm(96)
  df$frame <- x
  df
})

df_frames <- do.call(rbind, df_list)

# convert the data frame to multi-frame matrix
mat <- well_df_to_mat_frames(
  data = df_frames,
  values_col = "value",
  time_col = "frame"
)

head(mat[, 1:14])
#>           A01        A02        A03        A04        A05        A06        A07
#> 1 -0.49248211 -0.7996654  0.1844326 -0.2561536  1.0265133 -0.2859982  0.3586290
#> 2 -0.46363546  0.1595360  0.4634524 -0.4045746  0.6895686  0.5747917 -0.4565300
#> 3  0.02800806 -0.1388920  1.7883493  2.2458909 -0.9269583  0.1241107 -1.4833618
#> 4  0.19279652  1.1936065  1.5660579  0.9647351  1.1194541  0.5475243 -1.7128918
#> 5  0.54687357  3.5064287  0.5657981 -0.3352175  1.2802907 -0.1215788 -0.9465863
#> 6  0.77060832 -1.5865375 -0.7177842  0.6691126  1.0292328  0.4712009 -0.5783108
#>          A08        A09        A10         A11         A12        B01
#> 1 -0.7397960 -0.3620506  0.4986734 -0.03487612 -0.69042807 -1.8140603
#> 2  0.2452932  0.3397737 -0.6361876 -0.34111647  0.61346339  0.5575676
#> 3 -1.2477520  1.4945232  1.2275698  0.26418238 -0.87951638 -0.6371337
#> 4 -0.6168049  0.2591952 -0.2435477  0.84731986 -1.01752082 -0.3571054
#> 5  0.1717833 -0.4323345  0.7336416  0.19722168  0.06495437  1.7144867
#> 6  0.5420816  1.1270805  0.3009031 -0.46497856 -0.37079446  0.7343720
#>          B02
#> 1 -1.1186064
#> 2  0.8663748
#> 3 -1.2503462
#> 4  0.7603821
#> 5  0.3213963
#> 6 -0.3166930

# convert the matrix back to a dataframe
well_mat_frames_to_df(mat)
#> # A tibble: 960 × 5
#>    well    row   col  value frames
#>    <chr> <dbl> <dbl>  <dbl> <chr> 
#>  1 A01       1     1 -0.492 1     
#>  2 A02       1     2 -0.800 1     
#>  3 A03       1     3  0.184 1     
#>  4 A04       1     4 -0.256 1     
#>  5 A05       1     5  1.03  1     
#>  6 A06       1     6 -0.286 1     
#>  7 A07       1     7  0.359 1     
#>  8 A08       1     8 -0.740 1     
#>  9 A09       1     9 -0.362 1     
#> 10 A10       1    10  0.499 1     
#> # … with 950 more rows