Turn a Multi-Frame Matrix into a DataFrame

Takes a matrix where each row is a time point and each column is a well and converts it to a DataFrame containing a column for the well, a column for the row, a column for the column, a column for the time and a column for the values.

well_mat_frames_to_df(matrix, value_col = "value")

Arguments

matrix

multi-frame matrix to be converted.

value_col

Name to be assigned to the values column.

Value

a tibble

Examples

df_list <- lapply(1:10, function(x) {
  df <- wellr::well_plate()
  df$value <- rnorm(96)
  df$frame <- x
  df
})

df_frames <- do.call(rbind, df_list)

# convert the data frame to multi-frame matrix
mat <- well_df_to_mat_frames(
  data = df_frames,
  values_col = "value",
  time_col = "frame"
)

head(mat[, 1:14])
#>           A01        A02       A03        A04        A05         A06        A07
#> 1 -1.98188022  0.4010191 1.0936252  0.2016572 -2.6870403  0.07634781 -1.3636182
#> 2 -0.09233083  0.4987733 1.1310378 -0.6648330  0.2721987  0.20251025  0.9519429
#> 3 -1.46397394  0.4742682 0.8487191  0.7509885  1.8411537  0.93142382  0.4503291
#> 4  0.57701871  0.3994673 0.4656361  0.6753026 -0.3797288 -0.27077757 -0.6399890
#> 5  0.03902588  0.4882045 1.6852477  0.4957293 -0.4055078 -0.02288954 -1.2252467
#> 6  1.11924825 -1.0073503 0.9000788 -0.2954812 -0.9769861 -0.61410334  0.3342462
#>          A08        A09        A10          A11        A12         B01
#> 1  1.3350963  0.1518215 -0.6743018 -1.082997089 -0.7179322  2.31497489
#> 2  1.3019698  1.1912543  0.7432629  0.075006605  1.9174315  1.04120346
#> 3 -0.7506550 -0.5764367  0.9149543  1.127922702 -0.4132626  0.77274415
#> 4 -1.0375879  1.4993233  1.1675064  0.631233749 -0.2616410  1.13206484
#> 5 -0.3151919  0.8242082 -0.9171479  0.009110162  1.1363410 -0.89433451
#> 6 -0.9441064  0.2442029 -0.7324802 -0.026581758  0.9603215 -0.07447516
#>          B02
#> 1  0.9435254
#> 2  0.2784969
#> 3 -0.5568875
#> 4 -1.3667878
#> 5  0.3956234
#> 6 -2.0295057

# convert the matrix back to a dataframe
well_mat_frames_to_df(mat)
#> # A tibble: 960 × 5
#>    well    row   col   value frames
#>    <chr> <dbl> <dbl>   <dbl> <chr> 
#>  1 A01       1     1 -1.98   1     
#>  2 A02       1     2  0.401  1     
#>  3 A03       1     3  1.09   1     
#>  4 A04       1     4  0.202  1     
#>  5 A05       1     5 -2.69   1     
#>  6 A06       1     6  0.0763 1     
#>  7 A07       1     7 -1.36   1     
#>  8 A08       1     8  1.34   1     
#>  9 A09       1     9  0.152  1     
#> 10 A10       1    10 -0.674  1     
#> # … with 950 more rows